% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{All.Normality.Tests}
\alias{All.Normality.Tests}
\title{All.Normality.Tests}
\usage{
All.Normality.Tests(
  dt = NULL,
  YVars = NULL,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "600px",
  PlotWidth = "300px",
  SampleSize.ADT = NULL,
  Samples.ADT = 1,
  SampleSize.CVMT = NULL,
  Samples.CVMT = 1,
  SampleSize.KST = NULL,
  Samples.KST = 1,
  SampleSize.ST = NULL,
  Samples.ST = 1,
  SampleSize.JBT = NULL,
  Samples.JBT = 1,
  SampleSize.AT = NULL,
  Samples.AT = 1
)
}
\description{
All.Normality.Tests
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
