% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{Anderson.Darling.Test}
\alias{Anderson.Darling.Test}
\title{Anderson.Darling.Test}
\usage{
Anderson.Darling.Test(Vals, SampleSize = NULL, Samples = 1)
}
\arguments{
\item{Vals}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Anderson.Darling.Test
}
\examples{
\dontrun{
Vals <- qnorm(p = runif(100000))
Anderson.Darling.Test(Vals)
}

}
\references{
https://en.m.wikipedia.org/wiki/Anderson%E2%80%93Darling_test, https://real-statistics.com/non-parametric-tests/goodness-of-fit-tests/anderson-darling-test/
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\concept{Inference}
