% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{AnomalyDetection_Modal_Fun}
\alias{AnomalyDetection_Modal_Fun}
\title{AnomalyDetection_Modal_Fun}
\usage{
AnomalyDetection_Modal_Fun(
  id,
  AppWidth = 12L,
  IsolationForest_H2O_Features_Selected = NULL,
  IsolationForest_H2O_TrainData_Choices = NULL,
  IsolationForest_H2O_ValidationData_Choices = NULL,
  IsolationForest_H2O_TestData_Choices = NULL,
  IsolationForest_H2O_Threshold_Choices = NULL,
  IsolationForest_H2O_NTrees_Choices = NULL,
  IsolationForest_H2O_MaxDepth_Choices = NULL,
  IsolationForest_H2O_MinRows_Choices = NULL,
  IsolationForest_H2O_RowSampleRate_Choices = NULL,
  IsolationForest_H2O_ColSampleRate_Choices = NULL,
  IsolationForest_H2O_ColSampleRatePerLevel_Choices = NULL,
  IsolationForest_H2O_ColSampleRatePerTree_Choices = NULL,
  IsolationForest_H2O_TrainData_Selected = NULL,
  IsolationForest_H2O_ValidationData_Selected = NULL,
  IsolationForest_H2O_TestData_Selected = NULL,
  IsolationForest_H2O_Threshold_Selected = NULL,
  IsolationForest_H2O_NTrees_Selected = NULL,
  IsolationForest_H2O_MaxDepth_Selected = NULL,
  IsolationForest_H2O_MinRows_Selected = NULL,
  IsolationForest_H2O_RowSampleRate_Selected = NULL,
  IsolationForest_H2O_ColSampleRate_Selected = NULL,
  IsolationForest_H2O_ColSampleRatePerLevel_Selected = NULL,
  IsolationForest_H2O_ColSampleRatePerTree_Selected = NULL
)
}
\arguments{
\item{id}{= 'AD_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Anomaly Detection operations
}
\seealso{
Other FE: 
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Categorical_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFillRoll}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_Clustering}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{NLP_Modal_Fun}()},
\code{\link{Numeric_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
