% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{AzureBlob_Modal_Fun}
\alias{AzureBlob_Modal_Fun}
\title{AzureBlob_Modal_Fun}
\usage{
AzureBlob_Modal_Fun(
  id,
  AppWidth = 12L,
  AzureBlobStorageTabular_Choices = NULL,
  AzureBlobStorageRdata_Choices = NULL,
  AzureBlobStorageTabular_Selected = NULL,
  AzureBlobStorageRdata_Selected = NULL
)
}
\arguments{
\item{id}{= 'AzureBlobDMID'}

\item{AppWidth}{= 12L}
}
\description{
AzureBlob Environment
}
\seealso{
Other DM: 
\code{\link{DM_AzureBlob}()},
\code{\link{DM_Local}()},
\code{\link{DM_PostGRE}()},
\code{\link{DM_Session}()},
\code{\link{Local_Modal_Fun}()},
\code{\link{PostGRE_Modal_Fun}()},
\code{\link{Session_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DM}
