% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{BasicLoop.SingleTest}
\alias{BasicLoop.SingleTest}
\title{BasicLoop.SingleTest}
\usage{
BasicLoop.SingleTest(
  LM,
  Plan,
  ArgsList,
  DataList,
  CodeList,
  ModelID,
  VD,
  Test,
  Lift = 0,
  LiftMetric = "MAE",
  Algo = "catboost",
  DebugFC = FALSE
)
}
\arguments{
\item{LM}{LoopMetrics From Backtest.SingelLoop}

\item{Plan}{From Backtest.SingelLoop}

\item{ArgsList}{List}

\item{DataList}{From app}

\item{CodeList}{From app}

\item{ModelID}{From Panel Forecasting Common Args Section of Shiny.FC.Trainer}

\item{VD}{ValidationData}

\item{Test}{= Test}

\item{Lift}{= 0. If MAE with args on is greater than MAE with args off by at least MAE_MinLift then the feature will be kept on}

\item{LiftMetric}{= 'MAE' or 'RMSE'}

\item{Algo}{= 'catboost', 'xgboost', 'lightgbm'}

\item{DebugFC}{From app}
}
\description{
Test out various combinations of structure parameters to narrow the search space
}
\seealso{
Other FC: 
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
\keyword{internal}
