% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{BlankRow}
\alias{BlankRow}
\title{BlankRow}
\usage{
BlankRow(W = 12)
}
\arguments{
\item{W}{width of column}
}
\value{
Adds a row to your UI of width W
}
\description{
BlankRow add blank row with width w
}
\examples{
\dontrun{
Quantico:::BlankRow(12)
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
