% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{BuildBinary}
\alias{BuildBinary}
\title{BuildBinary}
\usage{
BuildBinary(Root = NULL)
}
\arguments{
\item{Root}{NULL will setwd to project root as defined in function}
}
\description{
Build package binary
}
\seealso{
Other Utilities: 
\code{\link{Install}()},
\code{\link{UpdateDocs}()}
}
\author{
Adrian Antico
}
\concept{Utilities}
