% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{BuildModelsButton}
\alias{BuildModelsButton}
\title{BuildModelsButton}
\usage{
BuildModelsButton(
  id = "MLBuildButton",
  Algo = "CatBoost",
  Style = "gradient",
  Color = "royal"
)
}
\arguments{
\item{id}{= 'MLBuildButton'}

\item{Algo}{= 'CatBoost'}

\item{Style}{= 'gradient'}

\item{Color}{= 'royal'}
}
\description{
BuildModelsButton
}
\seealso{
Other Modeling: 
\code{\link{FC_Arfima}()},
\code{\link{FC_CatBoostCARMAFUN}()},
\code{\link{FC_CatBoostCARMAID}()},
\code{\link{FC_CatBoostCARMA}()},
\code{\link{FC_ETS}()},
\code{\link{FC_LightGBMCARMAFUN}()},
\code{\link{FC_LightGBMCARMAID}()},
\code{\link{FC_LightGBMCARMA}()},
\code{\link{FC_NNet}()},
\code{\link{FC_SARIMA}()},
\code{\link{FC_TBATS}()},
\code{\link{FC_XGBoostCARMAFUN}()},
\code{\link{FC_XGBoostCARMAID}()},
\code{\link{FC_XGBoostCARMA}()},
\code{\link{ML_CatBoostReport}()},
\code{\link{ML_CatBoost}()},
\code{\link{ML_CausalMediation}()},
\code{\link{ML_GEE}()},
\code{\link{ML_H2O_DRF}()},
\code{\link{ML_H2O_GBM}()},
\code{\link{ML_H2O_GLM}()},
\code{\link{ML_H2O_HGLM}()},
\code{\link{ML_LightGBMReport}()},
\code{\link{ML_LightGBM}()},
\code{\link{ML_MixedEffects}()},
\code{\link{ML_Scoring}()},
\code{\link{ML_XGBoostReport}()},
\code{\link{ML_XGBoost}()}
}
\author{
Adrian Antico
}
\concept{Modeling}
