% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{CatBoost_Modal_Fun}
\alias{CatBoost_Modal_Fun}
\title{CatBoost_Modal_Fun}
\usage{
CatBoost_Modal_Fun(
  id,
  AppWidth = 12L,
  CatBoost_TargetColumnName_Selected = NULL,
  CatBoost_FeatureColNames_Selected = NULL,
  CatBoost_WeightsColumnName_Selected = NULL,
  CatBoost_TransformNumericColumns_Selected = NULL,
  CatBoost_grid_eval_metric_Selected = NULL,
  CatBoost_LossFunction_Selected = NULL,
  CatBoost_EvalMetric_Selected = NULL,
  CatBoost_TargetType_Choices = NULL,
  CatBoost_Runs_Choices = NULL,
  CatBoost_data_Choices = NULL,
  CatBoost_ValidationData_Choices = NULL,
  CatBoost_TestData_Choices = NULL,
  CatBoost_EncodeMethod_Choices = NULL,
  CatBoost_Methods_Choices = NULL,
  CatBoost_PassInGrid_Choices = NULL,
  CatBoost_NThreads_Choices = NULL,
  CatBoost_task_type_Choices = NULL,
  CatBoost_NumGPUs_Choices = NULL,
  CatBoost_TrainOnFull_Choices = NULL,
  CatBoost_Trees_Choices = NULL,
  CatBoost_Depth_Choices = NULL,
  CatBoost_LearningRate_Choices = NULL,
  CatBoost_L2_Leaf_Reg_Choices = NULL,
  CatBoost_model_size_reg_Choices = NULL,
  CatBoost_langevin_Choices = NULL,
  CatBoost_diffusion_temperature_Choices = NULL,
  CatBoost_RandomStrength_Choices = NULL,
  CatBoost_BorderCount_Choices = NULL,
  CatBoost_RSM_Choices = NULL,
  CatBoost_BootStrapType_Choices = NULL,
  CatBoost_GrowPolicy_Choices = NULL,
  CatBoost_feature_border_type_Choices = NULL,
  CatBoost_subsample_Choices = NULL,
  CatBoost_score_function_Choices = NULL,
  CatBoost_min_data_in_leaf_Choices = NULL,
  CatBoost_GridTune_Choices = NULL,
  CatBoost_MaxModelsInGrid_Choices = NULL,
  CatBoost_MaxRunsWithoutNewWinner_Choices = NULL,
  CatBoost_MaxRunMinutes_Choices = NULL,
  CatBoost_BaselineComparison_Choices = NULL,
  CatBoost_MetricPeriods_Choices = NULL,
  CatBoost_ModelID_Selected = NULL,
  CatBoost_TargetType_Selected = NULL,
  CatBoost_Runs_Selected = NULL,
  CatBoost_data_Selected = NULL,
  CatBoost_ValidationData_Selected = NULL,
  CatBoost_TestData_Selected = NULL,
  CatBoost_EncodeMethod_Selected = NULL,
  CatBoost_Methods_Selected = NULL,
  CatBoost_PassInGrid_Selected = NULL,
  CatBoost_NThreads_Selected = NULL,
  CatBoost_task_type_Selected = NULL,
  CatBoost_NumGPUs_Selected = NULL,
  CatBoost_TrainOnFull_Selected = NULL,
  CatBoost_Trees_Selected = NULL,
  CatBoost_Depth_Selected = NULL,
  CatBoost_LearningRate_Selected = NULL,
  CatBoost_L2_Leaf_Reg_Selected = NULL,
  CatBoost_model_size_reg_Selected = NULL,
  CatBoost_langevin_Selected = NULL,
  CatBoost_diffusion_temperature_Selected = NULL,
  CatBoost_RandomStrength_Selected = NULL,
  CatBoost_BorderCount_Selected = NULL,
  CatBoost_RSM_Selected = NULL,
  CatBoost_BootStrapType_Selected = NULL,
  CatBoost_GrowPolicy_Selected = NULL,
  CatBoost_feature_border_type_Selected = NULL,
  CatBoost_subsample_Selected = NULL,
  CatBoost_score_function_Selected = NULL,
  CatBoost_min_data_in_leaf_Selected = NULL,
  CatBoost_GridTune_Selected = NULL,
  CatBoost_MaxModelsInGrid_Selected = NULL,
  CatBoost_MaxRunsWithoutNewWinner_Selected = NULL,
  CatBoost_MaxRunMinutes_Selected = NULL,
  CatBoost_BaselineComparison_Selected = NULL,
  CatBoost_MetricPeriods_Selected = NULL,
  CatBoost_ClassWeights0_Selected = NULL,
  CatBoost_ClassWeights1_Selected = NULL
)
}
\arguments{
\item{id}{= 'CatBoost_ID'}

\item{AppWidth}{= 12L}
}
\description{
CatBoost Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
