% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{CatBoostCARMA_Modal_Fun}
\alias{CatBoostCARMA_Modal_Fun}
\title{CatBoostCARMA_Modal_Fun}
\usage{
CatBoostCARMA_Modal_Fun(
  id,
  AppWidth = 12L,
  CatBoostCARMA_TargetColumnName_Selected = NULL,
  CatBoostCARMA_DateColumnName_Selected = NULL,
  CatBoostCARMA_GroupVariables_Selected = NULL,
  CatBoostCARMA_ModelID_Selected = NULL,
  CatBoostCARMA_RunMode_Choices = NULL,
  CatBoostCARMA_ArgsList_Choices = NULL,
  CatBoostCARMA_TrainData_Choices = NULL,
  CatBoostCARMA_ValidationData_Choices = NULL,
  CatBoostCARMA_XREGS_Choices = NULL,
  CatBoostCARMA_TimeUnit_Choices = NULL,
  CatBoostCARMA_FCPeriods_Choices = NULL,
  CatBoostCARMA_EncodingMethod_Choices = NULL,
  CatBoostCARMA_ZeroPadSeries_Choices = NULL,
  CatBoostCARMA_Methods_Choices = NULL,
  CatBoostCARMA_Difference_Choices = NULL,
  CatBoostCARMA_TimeWeights_Choices = NULL,
  CatBoostCARMA_NonNegativePred_Choices = NULL,
  CatBoostCARMA_RoundPreds_Choices = NULL,
  CatBoostCARMA_TaskType_Choices = NULL,
  CatBoostCARMA_NumGPU_Choices = NULL,
  CatBoostCARMA_SaveModel_Choices = NULL,
  CatBoostCARMA_CalendarVariables_Choices = NULL,
  CatBoostCARMA_HolidayVariables_Choices = NULL,
  CatBoostCARMA_Lookback_Choices = NULL,
  CatBoostCARMA_TimeTrend_Choices = NULL,
  CatBoostCARMA_AnomalyDetection_HighThreshold_Choices = NULL,
  CatBoostCARMA_AnomalyDetection_LowThreshold_Choices = NULL,
  CatBoostCARMA_Lags_Choices = NULL,
  CatBoostCARMA_MovingAverages_Choices = NULL,
  CatBoostCARMA_MovingSD_Choices = NULL,
  CatBoostCARMA_MovingSkew_Choices = NULL,
  CatBoostCARMA_MovingKurt_Choices = NULL,
  CatBoostCARMA_MovingQuantiles_Choices = NULL,
  CatBoostCARMA_Quantiles_Selected_Choices = NULL,
  CatBoostCARMA_DataTruncate_Choices = NULL,
  CatBoostCARMA_NTrees_Choices = NULL,
  CatBoostCARMA_Depth_Choices = NULL,
  CatBoostCARMA_Langevin_Choices = NULL,
  CatBoostCARMA_DiffusionTemperature_Choices = NULL,
  CatBoostCARMA_SubSample_Choices = NULL,
  CatBoostCARMA_BootStrapType_Choices = NULL,
  CatBoostCARMA_L2_Leaf_Reg_Choices = NULL,
  CatBoostCARMA_LearningRate_Choices = NULL,
  CatBoostCARMA_ModelSizeReg_Choices = NULL,
  CatBoostCARMA_MinDataInLeaf_Choices = NULL,
  CatBoostCARMA_RandomStrength_Choices = NULL,
  CatBoostCARMA_RSM_Choices = NULL,
  CatBoostCARMA_BorderCount_Choices = NULL,
  CatBoostCARMA_GrowPolicy_Choices = NULL,
  CatBoostCARMA_FeatureBorderType_Choices = NULL,
  CatBoostCARMA_ScoreFunction_Choices = NULL,
  CatBoostCARMA_CrossEvalModelUpdate_Choices = NULL,
  CatBoostCARMA_CrossEvalValuesUpdate_Choices = NULL,
  CatBoostCARMA_CrossEvalModelUpdateFreq_Choices = NULL,
  CatBoostCARMA_CrossEvalValuesUpdateFreq_Choices = NULL,
  CatBoostCARMA_EvalMetric_Choices = NULL,
  CatBoostCARMA_EvalMetricValue_Choices = NULL,
  CatBoostCARMA_LossFunction_Choices = NULL,
  CatBoostCARMA_LossFunctionValue_Choices = NULL,
  CatBoostCARMA_RunMode_Selected = NULL,
  CatBoostCARMA_ArgsList_Selected = NULL,
  CatBoostCARMA_TrainData_Selected = NULL,
  CatBoostCARMA_ValidationData_Selected = NULL,
  CatBoostCARMA_XREGS_Selected = NULL,
  CatBoostCARMA_TimeUnit_Selected = NULL,
  CatBoostCARMA_FCPeriods_Selected = NULL,
  CatBoostCARMA_EncodingMethod_Selected = NULL,
  CatBoostCARMA_ZeroPadSeries_Selected = NULL,
  CatBoostCARMA_Methods_Selected = NULL,
  CatBoostCARMA_Difference_Selected = NULL,
  CatBoostCARMA_TimeWeights_Selected = NULL,
  CatBoostCARMA_NonNegativePred_Selected = NULL,
  CatBoostCARMA_RoundPreds_Selected = NULL,
  CatBoostCARMA_TaskType_Selected = NULL,
  CatBoostCARMA_NumGPU_Selected = NULL,
  CatBoostCARMA_SaveModel_Selected = NULL,
  CatBoostCARMA_CalendarVariables_Selected = NULL,
  CatBoostCARMA_HolidayVariables_Selected = NULL,
  CatBoostCARMA_Lookback_Selected = NULL,
  CatBoostCARMA_TimeTrend_Selected = NULL,
  CatBoostCARMA_AnomalyDetection_HighThreshold_Selected = NULL,
  CatBoostCARMA_AnomalyDetection_LowThreshold_Selected = NULL,
  CatBoostCARMA_Lags_Selected = NULL,
  CatBoostCARMA_MovingAverages_Selected = NULL,
  CatBoostCARMA_MovingSD_Selected = NULL,
  CatBoostCARMA_MovingSkew_Selected = NULL,
  CatBoostCARMA_MovingKurt_Selected = NULL,
  CatBoostCARMA_MovingQuantiles_Selected = NULL,
  CatBoostCARMA_Quantiles_Selected_Selected = NULL,
  CatBoostCARMA_DataTruncate_Selected = NULL,
  CatBoostCARMA_NTrees_Selected = NULL,
  CatBoostCARMA_Depth_Selected = NULL,
  CatBoostCARMA_Langevin_Selected = NULL,
  CatBoostCARMA_DiffusionTemperature_Selected = NULL,
  CatBoostCARMA_SubSample_Selected = NULL,
  CatBoostCARMA_BootStrapType_Selected = NULL,
  CatBoostCARMA_L2_Leaf_Reg_Selected = NULL,
  CatBoostCARMA_LearningRate_Selected = NULL,
  CatBoostCARMA_ModelSizeReg_Selected = NULL,
  CatBoostCARMA_MinDataInLeaf_Selected = NULL,
  CatBoostCARMA_RandomStrength_Selected = NULL,
  CatBoostCARMA_RSM_Selected = NULL,
  CatBoostCARMA_BorderCount_Selected = NULL,
  CatBoostCARMA_GrowPolicy_Selected = NULL,
  CatBoostCARMA_FeatureBorderType_Selected = NULL,
  CatBoostCARMA_ScoreFunction_Selected = NULL,
  CatBoostCARMA_CrossEvalModelUpdate_Selected = NULL,
  CatBoostCARMA_CrossEvalValuesUpdate_Selected = NULL,
  CatBoostCARMA_CrossEvalModelUpdateFreq_Selected = NULL,
  CatBoostCARMA_CrossEvalValuesUpdateFreq_Selected = NULL,
  CatBoostCARMA_EvalMetric_Selected = NULL,
  CatBoostCARMA_EvalMetricValue_Selected = NULL,
  CatBoostCARMA_LossFunction_Selected = NULL,
  CatBoostCARMA_LossFunctionValue_Selected = NULL
)
}
\arguments{
\item{id}{= 'CatBoostCARMA_ID'}

\item{AppWidth}{= 12L}
}
\description{
CatBoost CARMA Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
