% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Categorical_Modal_Fun}
\alias{Categorical_Modal_Fun}
\title{Categorical_Modal_Fun}
\usage{
Categorical_Modal_Fun(
  id,
  AppWidth = 12L,
  CategoricalEncoding_GroupVariables_Selected = NULL,
  CategoricalEncoding_TargetVariable_Selected = NULL,
  DummifyDT_Cols_Selected = NULL,
  CategoricalEncoding_TrainData_Choices = NULL,
  CategoricalEncoding_Method_Choices = NULL,
  CategoricalEncoding_ValidationData_Choices = NULL,
  CategoricalEncoding_TestData_Choices = NULL,
  DummifyDT_SelectData_Choices = NULL,
  DummifyDT_TopN_Choices = NULL,
  DummifyDT_KeepBaseCols_Choices = NULL,
  CategoricalEncoding_TrainData_Selected = NULL,
  CategoricalEncoding_Method_Selected = NULL,
  CategoricalEncoding_ValidationData_Selected = NULL,
  CategoricalEncoding_TestData_Selected = NULL,
  DummifyDT_SelectData_Selected = NULL,
  DummifyDT_TopN_Selected = NULL,
  DummifyDT_KeepBaseCols_Selected = NULL
)
}
\arguments{
\item{id}{= 'Categorical_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Categorical variables
}
\seealso{
Other FE: 
\code{\link{AnomalyDetection_Modal_Fun}()},
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFillRoll}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_Clustering}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{NLP_Modal_Fun}()},
\code{\link{Numeric_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
