% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{CausalMediation_Modal_Fun}
\alias{CausalMediation_Modal_Fun}
\title{CausalMediation_Modal_Fun}
\usage{
CausalMediation_Modal_Fun(
  id,
  AppWidth = 12L,
  CausalMediation_OutcomeTargetVariable_Selected = NULL,
  CausalMediation_TreatmentVariable_Selected = NULL,
  CausalMediation_MediatorVariable_Selected = NULL,
  CausalMediation_SurvivalEventVariable_Selected = NULL,
  CausalMediation_Covariates_Selected = NULL,
  CausalMediation_OM_MediatorCovariates_Selected = NULL,
  CausalMediation_OM_TreatmentCovariates_Selected = NULL,
  CausalMediation_MM_TreatmentCovariates_Selected = NULL,
  CausalMediation_data_Choices = NULL,
  CausalMediation_OutcomeTargetType_Choices = NULL,
  CausalMediation_MediatorTargetType_Choices = NULL,
  CausalMediation_TreatmentMediatorInteraction_Choices = NULL,
  CausalMediation_RemoveNA_Choices = NULL,
  CausalMediation_CaseControlSourceData_Choices = NULL,
  CausalMediation_Treated_ReferenceIndicator_Choices = NULL,
  CausalMediation_UnTreated_ReferenceIndicator_Choices = NULL,
  CausalMediation_Mediator_ControlDirectEffectLevel_Choices = NULL,
  CausalMediation_Covariate_NaturalDirectIndirect_Choices = NULL,
  CausalMediation_data_Selected = NULL,
  CausalMediation_OutcomeTargetType_Selected = NULL,
  CausalMediation_MediatorTargetType_Selected = NULL,
  CausalMediation_TreatmentMediatorInteraction_Selected = NULL,
  CausalMediation_RemoveNA_Selected = NULL,
  CausalMediation_CaseControlSourceData_Selected = NULL,
  CausalMediation_Treated_ReferenceIndicator_Selected = NULL,
  CausalMediation_UnTreated_ReferenceIndicator_Selected = NULL,
  CausalMediation_Mediator_ControlDirectEffectLevel_Selected = NULL,
  CausalMediation_Covariate_NaturalDirectIndirect_Selected = NULL
)
}
\arguments{
\item{id}{= 'CausalMediation_ID'}

\item{AppWidth}{= 12L}
}
\description{
Causal Mediation Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
