% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{ChiSq.Test}
\alias{ChiSq.Test}
\title{ChiSq.Test}
\usage{
ChiSq.Test(
  dt = NULL,
  Variable1 = NULL,
  Variable2 = NULL,
  ConfidenceLevel = 0.95,
  Alternative = "two.sided",
  SampleSize = NULL,
  Samples = 1,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "850px",
  PlotWidth = "1450px"
)
}
\arguments{
\item{dt}{data.table}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}

\item{Variables1}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}

\item{Variables2}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}
}
\description{
ChiSq.Test
}
\examples{
\dontrun{
Variable1 = "Customer"
dt = data.table::fread(file.choose())
Variable2 = "Brand"
SampleSize = 500
Samples = 50
EchartsTheme = "macarons"
TextColor = "black"
PlotHeight = "850px"
PlotWidth = "1450px"
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\concept{Inference}
