% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{ChiSquareTest_Modal_Fun}
\alias{ChiSquareTest_Modal_Fun}
\title{ChiSquareTest_Modal_Fun}
\usage{
ChiSquareTest_Modal_Fun(
  id,
  AppWidth = 12L,
  ChiSquareTest_Variable1_Selected = NULL,
  ChiSquareTest_Variable2_Selected = NULL,
  ChiSquareTest_SelectData_Choices = NULL,
  ChiSquareTest_Alternative_Choices = NULL,
  ChiSquareTest_SelectData_Selected = NULL,
  ChiSquareTest_InferenceID_Selected = NULL,
  ChiSquareTest_Alternative_Selected = NULL,
  ChiSquareTest_ConfidenceLevel_Selected = NULL,
  ChiSquareTest_Samples_Selected = NULL,
  ChiSquareTest_SampleSize_Selected = NULL
)
}
\arguments{
\item{id}{= 'ChiSquareTest_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Inference operations
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{Inference}
