% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUtils.R
\name{CleanVector}
\alias{CleanVector}
\title{DataTable}
\usage{
CleanVector(x)
}
\arguments{
\item{x}{vector}
}
\description{
Fully loaded DT::datatable() with args prefilled
}
\seealso{
Other Utils: 
\code{\link{DataTable}()},
\code{\link{IntraSessionDefaults}()},
\code{\link{Shiny.CodePrint.Collect}()},
\code{\link{Shiny.CodePrint.OrganizeCode}()},
\code{\link{Shiny.Utils.CachePath}()},
\code{\link{Utils.Code.Lines}()},
\code{\link{Utils.LinuxRunSH}()},
\code{\link{Utils.Nest}()}
}
\author{
Adrian Antico
}
\concept{Utils}
