% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomePage.R
\name{CodePrintPage}
\alias{CodePrintPage}
\title{CodePrintPage}
\usage{
CodePrintPage(id, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotPanels'}

\item{AppWidth}{= 12L}

\item{Page}{numeric}

\item{NumPlotsAvailable}{numeric}

\item{DragulaChoices}{character}
}
\description{
Panels
}
\seealso{
Other Plotting: 
\code{\link{HomePage}()},
\code{\link{InferencePage}()},
\code{\link{SideBarUI}()}
}
\author{
Adrian Antico
}
\concept{Plotting}
