% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Columns_Modal_Fun}
\alias{Columns_Modal_Fun}
\title{Columns_Modal_Fun}
\usage{
Columns_Modal_Fun(
  id,
  AppWidth = 12L,
  TypeCast_Numeric_Selected = NULL,
  TypeCast_Integer_Selected = NULL,
  TypeCast_Character_Selected = NULL,
  TypeCast_Factor_Selected = NULL,
  TypeCast_Logical_Selected = NULL,
  TypeCast_Date_Selected = NULL,
  TypeCast_ExcelDate_Selected = NULL,
  TypeCast_Posix_Selected = NULL,
  RenameColumns_RenameColumn_Selected = NULL,
  TimeTrendColumn_DateColumn_Selected = NULL,
  TimeTrendColumn_GroupVars_Selected = NULL,
  ConcatColumns_Selected = NULL,
  TypeCast_SelectData_Choices = NULL,
  RenameColumns_SelectData_Choices = NULL,
  TimeTrendColumn_SelectData_Choices = NULL,
  TimeTrendColumn_TimeTrendOrder_Choices = NULL,
  ConcatColumns_SelectData_Choices = NULL,
  TypeCast_SelectData_Selected = NULL,
  RenameColumns_SelectData_Selected = NULL,
  RenameColumns_NewName_Selected = NULL,
  TimeTrendColumn_SelectData_Selected = NULL,
  TimeTrendColumn_TimeTrendOrder_Selected = NULL,
  TimeTrendColumn_NewName_Selected = NULL,
  ConcatColumns_SelectData_Selected = NULL
)
}
\arguments{
\item{id}{= 'Columns_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Melt and cast data
}
\seealso{
Other DW: 
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
