% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Correlation_Modal_Fun}
\alias{Correlation_Modal_Fun}
\title{Correlation_Modal_Fun}
\usage{
Correlation_Modal_Fun(
  id = "Inference_CorrelationID",
  AppWidth = 12L,
  Correlation_CorrVars_Selected = NULL,
  Correlation_DateVar_Selected = NULL,
  Correlation_SelectData_Choices = NULL,
  Correlation_P_Adjust_Choices = NULL,
  Correlation_Bayesian_Choices = NULL,
  Correlation_Bayesian_Prior_Choices = NULL,
  Correlation_MultiLevel_Choices = NULL,
  Correlation_Include_Factors_Choices = NULL,
  Correlation_Partial_Choices = NULL,
  Correlation_Partial_Bayesian_Choices = NULL,
  Correlation_SelectData_Selected = NULL,
  Correlation_InferenceID_Selected = NULL,
  Correlation_SampleSize_Selected = NULL,
  Correlation_P_Adjust_Selected = NULL,
  Correlation_Bayesian_Selected = NULL,
  Correlation_Bayesian_Prior_Selected = NULL,
  Correlation_MultiLevel_Selected = NULL,
  Correlation_Include_Factors_Selected = NULL,
  Correlation_Partial_Selected = NULL,
  Correlation_Partial_Bayesian_Selected = NULL
)
}
\arguments{
\item{id}{= 'Correlation_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Inference operations
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{Shapiro.Test}()}
}
\author{
Adrian Antico
}
\concept{Inference}
