% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Cramer.Von.Mises.Test}
\alias{Cramer.Von.Mises.Test}
\title{Cramer.Von.Mises.Test}
\usage{
Cramer.Von.Mises.Test(Vals, SampleSize = NULL, Samples = 1)
}
\arguments{
\item{Vals}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Cramer.Von.Mises.Test
}
\examples{
\dontrun{
Vals <- qnorm(p = runif(100000))
Cramer.Von.Mises.Test(Vals)
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Correlation.Test}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
