% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{DataFilters}
\alias{DataFilters}
\title{DataFilters}
\usage{
DataFilters(
  id = "DataFiltersContents",
  PlotNumber = 1,
  H3Color = parent.frame()$H3Color,
  H4Color = parent.frame()$H4Color,
  Right = FALSE,
  Animate = TRUE,
  Status = "custom"
)
}
\arguments{
\item{id}{= 'DataFilters'}

\item{PlotNumber}{= 1}

\item{H3Color}{= H3Color}

\item{H4Color}{= H4Color}

\item{Right}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}
}
\description{
Display a dropdown button with filter variables and other filter info needed for server
}
\seealso{
Other Plotting Modals: 
\code{\link{Plots.3D.Modal.NoFacet}()},
\code{\link{Plots.3D.Modal}()},
\code{\link{Plots.3D.NoGroup.Modal}()},
\code{\link{Plots.3D.Target.NoGroup.Modal}()},
\code{\link{Plots.ACF.Modal}()},
\code{\link{Plots.AggRel.Modal2}()},
\code{\link{Plots.AggRel.Modal}()},
\code{\link{Plots.AggRel.NoFacet.Modal}()},
\code{\link{Plots.CorrMatrix.Modal}()},
\code{\link{Plots.Distribution.Modal}()},
\code{\link{Plots.Distribution.XY.NoFacet.Modal}()},
\code{\link{Plots.Distribution.Y.NoFacet.Modal}()},
\code{\link{Plots.Eval.Modal}()},
\code{\link{Plots.Eval.ShapImp.Modal}()},
\code{\link{Plots.Eval.VarImp.Modal}()},
\code{\link{Plots.Eval.XY.Modal}()},
\code{\link{Plots.Eval.XYZ.NoGroup.Modal}()},
\code{\link{Plots.Eval.ZY.Modal}()},
\code{\link{Plots.Eval.ZY.NoFacet.Modal}()},
\code{\link{Plots.ParallelPlot.Modal}()},
\code{\link{Plots.ProbabilityPlot.Modal}()},
\code{\link{Plots.Radar.Modal}()},
\code{\link{Plots.WordCloud.Modal}()}
}
\author{
Adrian Antico
}
\concept{Plotting Modals}
