% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{DataPanels}
\alias{DataPanels}
\title{DataPanels}
\usage{
DataPanels(id, Page, AppWidth = 12L, DL = NULL)
}
\arguments{
\item{id}{= 'DataPanels'}

\item{Page}{numeric}

\item{AppWidth}{= 12L}

\item{DL}{character}
}
\description{
Panels
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
