% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{DataSets_Modal_Fun}
\alias{DataSets_Modal_Fun}
\title{DataSets_Modal_Fun}
\usage{
DataSets_Modal_Fun(
  id,
  AppWidth = 12L,
  AutoDataPartition_StratifyColumnNames_Selected = NULL,
  AutoDataPartition_TimeColumnName_Selected = NULL,
  SortData_SortColumns_Selected = NULL,
  ModelDataPrep_IgnoreCols_Selected = NULL,
  AutoDataPartition_SelectData_Choices = NULL,
  AutoDataPartition_NumDataSets_Choices = NULL,
  AutoDataPartition_Ratios_Train_Choices = NULL,
  AutoDataPartition_Ratios_Validation_Choices = NULL,
  AutoDataPartition_Ratios_Test_Choices = NULL,
  AutoDataPartition_PartitionType_Choices = NULL,
  SortData_SelectData_Choices = NULL,
  SortData_SortOrder_Choices = NULL,
  ModelDataPrep_SelectData_Choices = NULL,
  ModelDataPrep_CharToFactor_Choices = NULL,
  ModelDataPrep_FactorToChar_Choices = NULL,
  ModelDataPrep_DateToChar_Choices = NULL,
  ModelDataPrep_IDateConversion_Choices = NULL,
  ModelDataPrep_RemoveDates_Choices = NULL,
  ModelDataPrep_IntToNumeric_Choices = NULL,
  ModelDataPrep_LogicalToBinary_Choices = NULL,
  RemoveData_Datasets_Choices = NULL,
  AutoDataPartition_SelectData_Selected = NULL,
  AutoDataPartition_NumDataSets_Selected = NULL,
  AutoDataPartition_Ratios_Train_Selected = NULL,
  AutoDataPartition_Ratios_Validation_Selected = NULL,
  AutoDataPartition_Ratios_Test_Selected = NULL,
  AutoDataPartition_PartitionType_Selected = NULL,
  SortData_SelectData_Selected = NULL,
  SortData_SortOrder_Selected = NULL,
  ModelDataPrep_SelectData_Selected = NULL,
  ModelDataPrep_CharToFactor_Selected = NULL,
  ModelDataPrep_FactorToChar_Selected = NULL,
  ModelDataPrep_DateToChar_Selected = NULL,
  ModelDataPrep_IDateConversion_Selected = NULL,
  ModelDataPrep_RemoveDates_Selected = NULL,
  ModelDataPrep_IntToNumeric_Selected = NULL,
  ModelDataPrep_LogicalToBinary_Selected = NULL,
  RemoveData_Datasets_Selected = NULL
)
}
\arguments{
\item{id}{= 'DataSets_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
DataSet operations
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
