% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.BenchmarkData}
\alias{DM.BenchmarkData}
\title{DM.BenchmarkData}
\usage{
DM.BenchmarkData(
  N = 10000000,
  Levels = 1000000,
  NAs = -1L,
  Sort = TRUE,
  Push2PostGRE = FALSE,
  Table = "H2O_BM_10m_1m",
  Host = "localhost",
  CloseConnection = FALSE,
  DBName = "AutoQuant",
  User = "postgres",
  Port = 5432,
  Password = "Aa1028#@"
)
}
\arguments{
\item{N}{= 10000000,}

\item{Levels}{= 1000000}

\item{NAs}{= -1L}

\item{Sort}{= TRUE}

\item{Push2PostGRE}{= FALSE}

\item{Table}{= 'H2O_BM_10m_1m'}

\item{Host}{= 'localhost'}

\item{CloseConnection}{= FALSE}

\item{DBName}{= 'AutoQuant'}

\item{User}{= 'postgres'}

\item{Port}{= 5432}

\item{Password}{= 'Aa1028#}
}
\description{
Data generator from the h2o benchmark study
}
