% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.DataListUpdate}
\alias{DM.DataListUpdate}
\title{DM.DataListUpdate}
\usage{
DM.DataListUpdate(dl, dn, Sample = TRUE, SampleSize = 1000)
}
\arguments{
\item{dl}{= DataList from App}

\item{dn}{= Dataset name stored inside DataList}

\item{Sample}{Create a sample}

\item{SampleSize}{default 1000 rows}
}
\value{
list with names 'data', 'cols', 'rows', 'colnames'
}
\description{
Store useful info
}
\keyword{internal}
