% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgAppend}
\alias{DM.pgAppend}
\title{DM.pgAppend}
\usage{
DM.pgAppend(
  data = NULL,
  DataBase = NULL,
  Table = NULL,
  Append = FALSE,
  Host = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{data}{Source data.table}

\item{DataBase}{dbname}

\item{Append}{Set to TRUE to append data, FALSE to overwrite data}

\item{Host}{host}

\item{User}{If Connection is NULL then this must be supplied. user}

\item{Port}{If Connection is NULL then this must be supplied. port}

\item{Password}{If Connection is NULL then this must be supplied. password}
}
\description{
PostGRE DM.pgAppend get data from a database table
}
\examples{
\dontrun{
Quantico::DM.pgAppend(
  data = data,
  Table = 'somename',
  Append = FALSE,
  Host = 'localhost',
  DataBase = 'AutoQuant',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')

# data = data
# Host = 'localhost'
# DataBase = 'Testing'
# User = 'postgres'
# Port = 5432
# Password = 'Aa...'
}

}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
