% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgCreateDB}
\alias{DM.pgCreateDB}
\title{DM.pgCreateDB}
\usage{
DM.pgCreateDB(
  DataBase = NULL,
  Host = "localhost",
  Port = 5432,
  User = "postgres",
  Password = ""
)
}
\arguments{
\item{DataBase}{See args from related functions}

\item{Host}{See args from related functions}

\item{Port}{See args from related functions}

\item{User}{See args from related functions}

\item{Password}{See args from related functions}
}
\description{
PostGRE DM.pgCreateDB will create a database with a name supplied by user
}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
