% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgCreateTable}
\alias{DM.pgCreateTable}
\title{DM.pgCreateTable}
\usage{
DM.pgCreateTable(
  data = NULL,
  DataBase = NULL,
  Schema = NULL,
  Table = NULL,
  Host = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{data}{Source data.table. If you supply a Schema, data will be ignored.}

\item{DataBase}{If Connection is NULL then this must be supplied. database name}

\item{Schema}{Optional. Advised to use if type inference is fuzzy}

\item{Table}{Name of table you want created}

\item{Host}{If Connection is NULL then this must be supplied. host name}

\item{User}{If Connection is NULL then this must be supplied. user name}

\item{Port}{If Connection is NULL then this must be supplied. port name}

\item{Password}{user password}
}
\description{
DM.pgCreateTable get data from a database table
}
\examples{
\dontrun{
DataMuse::DM.pgCreateTable(
  data,
  DataBase = 'Testing',
  Schema = NULL,
  Table = NULL,
  Host = 'localhost',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')
}

}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
