% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgListDatabases}
\alias{DM.pgListDatabases}
\title{DM.pgListDatabases}
\usage{
DM.pgListDatabases(
  Host = "localhost",
  Port = 5432,
  User = "postgres",
  Password = ""
)
}
\arguments{
\item{Host}{See args from related functions}

\item{Port}{See args from related functions}

\item{User}{See args from related functions}

\item{Password}{See args from related functions}
}
\description{
DM.pgListDatabases list of available databases
}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
