% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgQuery}
\alias{DM.pgQuery}
\title{DM.pgQuery}
\usage{
DM.pgQuery(
  Host = NULL,
  Query = NULL,
  DataBase = NULL,
  SELECT = "*",
  AggStat = "AVG",
  FROM = NULL,
  GroupBy = NULL,
  SamplePercent = 1,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{Host}{If Connection is NULL then this must be supplied. host}

\item{Query}{SQL Statement in quotes}

\item{DataBase}{From App}

\item{SELECT}{From App}

\item{AggStat}{From App}

\item{FROM}{From App}

\item{GroupBy}{From App}

\item{SamplePercent}{From App}

\item{User}{If Connection is NULL then this must be supplied. user}

\item{Port}{If Connection is NULL then this must be supplied. port}

\item{Password}{If Connection is NULL then this must be supplied. password}

\item{DBName}{If Connection is NULL then this must be supplied. dbname}
}
\description{
PostGRE DM.pgQuery get data from a database table
}
\examples{
\dontrun{

PlottingExample <- qs::qload(file = file.choose())
Columns <- c(
  "DATE_ISO",
  "ARTICLE",
  "BRAND",
  "CUSTOMER_COD_char",
  "CHILLED_Margin_PerDay",
  "CHILLED_Liters_PerDay",
  "CHILLED_Units_PerDay")
data <- DataMuse::DM.pgQuery(
  Query = NULL,
  DataBase = "KompsProcessed",
  SELECT = Columns,
  AggStat = "AVG",
  FROM = "POS_Processed_Long_Daily_backward",
  GroupBy = NULL,
  SamplePercent = 1,
  Host = 'localhost',
  User = 'postgres',
  Port = 5432,
  Password = "Aa...")
}

}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
