% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgRemoveCreateAppend}
\alias{DM.pgRemoveCreateAppend}
\title{DM.pgRemoveCreateAppend}
\usage{
DM.pgRemoveCreateAppend(
  data = NULL,
  DataBase = NULL,
  Table = NULL,
  Schema = NULL,
  Host = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{data}{See args from related functions}

\item{DataBase}{See args from related functions}

\item{Table}{See args from related functions}

\item{Schema}{See args from related functions}

\item{Host}{See args from related functions}

\item{User}{See args from related functions}

\item{Port}{See args from related functions}

\item{Password}{See args from related functions}
}
\description{
DM.pgRemoveCreateAppend will DROP the table specified
}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
