% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgRemoveTable}
\alias{DM.pgRemoveTable}
\title{DM.pgRemoveTable}
\usage{
DM.pgRemoveTable(
  DataBase = NULL,
  Table = NULL,
  Host = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{DataBase}{If Connection is NULL then this must be supplied. database name}

\item{Table}{Name of table you want created}

\item{Host}{If Connection is NULL then this must be supplied. Host name}

\item{User}{If Connection is NULL then this must be supplied. user name}

\item{Port}{If Connection is NULL then this must be supplied. port name}

\item{Password}{If Connection is NULL then this must be supplied. user password}
}
\description{
PostGRE DM.pgRemoveTable will DROP the table specified
}
\examples{
\dontrun{
Rappure::DM.pgRemoveTable(
  Table = 'static_data',
  Host = 'localhost',
  DataBase = 'Testing',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')

# Host = 'localhost'
# Table = 'static_data'
# Connection = NULL
# DataBase = 'Testing'
# User = 'postgres'
# Port = 5432
# Password = 'Aa...'
}

}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
