% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.pgTableColnames}
\alias{DM.pgTableColnames}
\title{DM.pgTableColnames}
\usage{
DM.pgTableColnames(
  Host = NULL,
  DataBase = NULL,
  Table = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{Host}{See args from related functions}

\item{DataBase}{See args from related functions}

\item{Table}{Name of postgres table}

\item{User}{See args from related functions}

\item{Port}{See args from related functions}

\item{Password}{See args from related functions}

\item{CloseConnection}{See args from related functions}
}
\value{
A character vector of names. Exactly like names R base function for a data.frame
}
\description{
PostGRE DM.pgTableColnames will list all column names from a table
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
