% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.ssClearTable}
\alias{DM.ssClearTable}
\title{DM.ssClearTable}
\usage{
DM.ssClearTable(DataBaseName, ServerName, SQLTableName = "", Errors = TRUE)
}
\arguments{
\item{DataBaseName}{DB Name from SS}

\item{ServerName}{Server name from SS}

\item{SQLTableName}{The SQL statement you want to run}

\item{Errors}{Set to TRUE to halt, FALSE to return -1 in cases of errors}
}
\description{
DM.ssClearTable remove all rows of a table
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
