% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.ssCreateTable}
\alias{DM.ssCreateTable}
\title{DM.ssCreateTable}
\usage{
DM.ssCreateTable(
  DataToPush,
  DataBaseName,
  ServerName,
  SQLTableName = "",
  RowNames = FALSE,
  ColNames = TRUE,
  AppendData = FALSE,
  AddPK = TRUE,
  Safer = TRUE
)
}
\arguments{
\item{DataToPush}{data to be sent to warehouse}

\item{DataBaseName}{DB Name from SS}

\item{ServerName}{Server name from SS}

\item{SQLTableName}{The SQL statement you want to run}

\item{RowNames}{either logical or character. If logical, save the row names as the first column rownames in the table? If character, the column name under which to save the rownames.}

\item{ColNames}{logical: save column names as the first row of table?}

\item{AppendData}{logical. Should data be appended to an existing table?}

\item{AddPK}{logical. Should rownames (if included) be specified as a primary key?}

\item{Safer}{logical. If true, create a non-existing table but only allow appends to an existing table. If false, allow sqlSave to attempt to delete all the rows of an existing table, or to drop it.}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}
}
\description{
DM.ssCreateTable create a database table or append data to an existing table
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
