% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.ssDictionaries}
\alias{DM.ssDictionaries}
\title{AutoDataDictionaries}
\usage{
DM.ssDictionaries(
  Type = "sqlserver",
  DBConnection,
  DDType = 1L,
  Query = NULL,
  ASIS = FALSE,
  CloseChannel = TRUE
)
}
\arguments{
\item{Type}{= "sqlserver" is currently the only system supported}

\item{DBConnection}{This is a RODBC connection object for sql server}

\item{DDType}{Select from 1 - 6 based on this article}

\item{Query}{Supply a query}

\item{ASIS}{Set to TRUE to pull in values without coercing types}

\item{CloseChannel}{Set to TRUE to disconnect}
}
\description{
AutoDataDictionaries is a function to return data dictionary data in table form
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
