% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.ssDropDB}
\alias{DM.ssDropDB}
\title{DM.ssDropDB}
\usage{
DM.ssDropDB(DataBaseName = "Muse", ServerName = "DESKTOP-AIC2MUB")
}
\arguments{
\item{DataBaseName}{DB Name from SS}

\item{ServerName}{Server name from SS}
}
\description{
PostGRE DM.ssDropDB will create a database with a name supplied by user
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssQuery}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
