% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{DM.ssQuery}
\alias{DM.ssQuery}
\title{DM.ssQuery}
\usage{
DM.ssQuery(DataBaseName, ServerName, Query, ASIS = FALSE, RowsPerBatch = 1024)
}
\arguments{
\item{DataBaseName}{DB Name from SS}

\item{ServerName}{Server name from SS}

\item{Query}{The SQL statement you want to run}

\item{ASIS}{Auto column typing}

\item{RowsPerBatch}{Rows default is 1024}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}
}
\description{
DM.ssQuery get data from a ss database table
}
\seealso{
Other Database: 
\code{\link{DM.SS.DBConnection}()},
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()},
\code{\link{DM.ssAppend}()},
\code{\link{DM.ssClearTable}()},
\code{\link{DM.ssCreateDB}()},
\code{\link{DM.ssCreateTable}()},
\code{\link{DM.ssDictionaries}()},
\code{\link{DM.ssDropDB}()},
\code{\link{DM.ssRemoveTable}()},
\code{\link{QueryBuilder}()}
}
\author{
Adrian Antico
}
\concept{Database}
