% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{DM_Local}
\alias{DM_Local}
\title{DM_Local}
\usage{
DM_Local(
  id = "DM_Local",
  Style = "gradient",
  Color = "royal",
  Align = "center",
  DropDownRight = FALSE,
  Animate = TRUE,
  Status = "custom",
  H3Color = "white"
)
}
\arguments{
\item{id}{= input}

\item{Style}{= 'gradient',}

\item{Color}{= 'royal',}

\item{Align}{= 'center'}

\item{DropDownRight}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}

\item{H3Color}{= 'white'}
}
\description{
Import and export from your local environment
}
\seealso{
Other DM: 
\code{\link{AzureBlob_Modal_Fun}()},
\code{\link{DM_AzureBlob}()},
\code{\link{DM_PostGRE}()},
\code{\link{DM_Session}()},
\code{\link{Local_Modal_Fun}()},
\code{\link{PostGRE_Modal_Fun}()},
\code{\link{Session_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DM}
