% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{DW_ShrinkData}
\alias{DW_ShrinkData}
\title{DW_ShrinkData}
\usage{
DW_ShrinkData(
  id = "DW_ShrinkData",
  Style = "gradient",
  Color = "royal",
  Align = "center",
  DropDownRight = FALSE,
  Animate = TRUE,
  Status = "custom",
  H3Color = "white"
)
}
\arguments{
\item{id}{= 'DW_ShrinkData'}

\item{Style}{= 'gradient',}

\item{Color}{= 'royal',}

\item{Align}{= 'center'}

\item{DropDownRight}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}

\item{H3Color}{= 'white'}
}
\description{
Subset, Aggregate, and Sample data
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
