% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{EDAPanels}
\alias{EDAPanels}
\title{EDAPanels}
\usage{
EDAPanels(id, Page, AppWidth = 12L, DL = NULL)
}
\arguments{
\item{id}{= 'EDAPanels'}

\item{Page}{numeric}

\item{AppWidth}{= 12L}

\item{DL}{numeric}
}
\description{
Panels
}
\seealso{
Other Output Panels: 
\code{\link{FCPanels}()},
\code{\link{InferencePanels}()},
\code{\link{MLPanels}()},
\code{\link{PlotDropDownContents}()},
\code{\link{PlotPanels}()}
}
\author{
Adrian Antico
}
\concept{Output Panels}
