% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{EffectsInterval}
\alias{EffectsInterval}
\title{EffectsInterval}
\usage{
EffectsInterval(Model)
}
\arguments{
\item{Model}{lmer and brms models}
}
\description{
Random Effects with confidence endpoints
}
\seealso{
Other Mixed Effects: 
\code{\link{FastLM}()},
\code{\link{LMER}()},
\code{\link{MEOW}()},
\code{\link{PredictionIntervals}()},
\code{\link{REsim}()}
}
\author{
Adrian Antico
}
\concept{Mixed Effects}
