% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{FastLM}
\alias{FastLM}
\title{FastLM}
\usage{
FastLM(data, yvar = NULL, xvar = NULL, Constrained = TRUE)
}
\arguments{
\item{data}{data.table}

\item{yvar}{target variable name}

\item{xvar}{x variable names}
}
\description{
LM with coefficients greater than or equal to zero if Constrained is TRUE. Otherwise, a fast lm coefficient set is returned from armadillo
}
\seealso{
Other Mixed Effects: 
\code{\link{EffectsInterval}()},
\code{\link{LMER}()},
\code{\link{MEOW}()},
\code{\link{PredictionIntervals}()},
\code{\link{REsim}()}
}
\author{
Adrian Antico
}
\concept{Mixed Effects}
