% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{FC.Backtest.CalendarGroups}
\alias{FC.Backtest.CalendarGroups}
\title{FC.Backtest.CalendarGroups}
\usage{
FC.Backtest.CalendarGroups(ArgsList)
}
\arguments{
\item{ArgsList}{List}
}
\description{
Forecasting TimeUnit expansion for backtesting eval rollups
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
