% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFE.R
\name{FE.Impute.TimeSeriesFillRoll}
\alias{FE.Impute.TimeSeriesFillRoll}
\title{FE.Impute.TimeSeriesFillRoll}
\usage{
FE.Impute.TimeSeriesFillRoll(
  input,
  output,
  session,
  DataList,
  CodeList,
  CacheDir = NULL,
  CacheName = "data",
  Debug = FALSE
)
}
\arguments{
\item{input}{passthrough}

\item{output}{passthrough}

\item{session}{passthrough}

\item{DataList}{passthrough}

\item{CodeList}{passthrough}

\item{CacheDir}{passthrough}

\item{CacheName}{passthrough}

\item{Debug}{passthrough}
}
\description{
Back and front fill a panel data set with irregular start and end dates
}
\seealso{
Other FE: 
\code{\link{AnomalyDetection_Modal_Fun}()},
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Categorical_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_Clustering}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{NLP_Modal_Fun}()},
\code{\link{Numeric_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
