% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{FE_Clustering}
\alias{FE_Clustering}
\title{FE_Clustering}
\usage{
FE_Clustering(
  id = "ClusteringID",
  Style = "gradient",
  Color = "royal",
  Align = "center",
  DropDownRight = FALSE,
  Animate = TRUE,
  Status = "custom",
  H3Color = "white"
)
}
\arguments{
\item{id}{= 'ModelBased_Modal'}

\item{Style}{= 'gradient',}

\item{Color}{= 'royal',}

\item{Align}{= 'center'}

\item{DropDownRight}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}

\item{H3Color}{= 'white'}
}
\description{
Unsupervised learning
}
\seealso{
Other FE: 
\code{\link{AnomalyDetection_Modal_Fun}()},
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Categorical_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFillRoll}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{NLP_Modal_Fun}()},
\code{\link{Numeric_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
