% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{FilterLogicData}
\alias{FilterLogicData}
\title{FilterLogicData}
\usage{
FilterLogicData(
  data1,
  FilterLogic = NULL,
  FilterVariable = NULL,
  FilterValue = NULL,
  FilterValue2 = NULL,
  Debug = FALSE
)
}
\arguments{
\item{data1}{data.table}

\item{FilterLogic}{passthrough}

\item{FilterVariable}{passthrough}

\item{FilterValue}{passthrough}

\item{FilterValue2}{passthrough}

\item{Debug}{passthrough}
}
\description{
FilterLogicData
}
