% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{FTest_Modal_Fun}
\alias{FTest_Modal_Fun}
\title{FTest_Modal_Fun}
\usage{
FTest_Modal_Fun(
  id,
  AppWidth = 12L,
  FTest_Variable1_Selected = NULL,
  FTest_Variable2_Selected = NULL,
  FTest_SelectData_Choices = NULL,
  FTest_Alternative_Choices = NULL,
  FTest_SelectData_Selected = NULL,
  FTest_InferenceID_Selected = NULL,
  FTest_RatioVariances_Selected = NULL,
  FTest_Alternative_Selected = NULL,
  FTest_ConfidenceLevel_Selected = NULL,
  FTest_Samples_Selected = NULL,
  FTest_SampleSize_Selected = NULL
)
}
\arguments{
\item{id}{= 'FTest_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Inference operations
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{Inference}
