% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{GenerateEvaluationMetrics}
\alias{GenerateEvaluationMetrics}
\title{GenerateEvaluationMetrics}
\usage{
GenerateEvaluationMetrics(
  EvalData = NULL,
  TargetName = NULL,
  DateName = NULL,
  GroupNames = NULL
)
}
\arguments{
\item{EvalData}{Source data in shiny app}

\item{TargetName}{Target variable name}

\item{DateName}{Date variable name}

\item{GroupNames}{Group variable names}
}
\value{
PreparePlotData object for server.R to
}
\description{
GenerateEvaluationMetrics calculates evaluation metrics for out of sample forecast and evaluation data
}
\examples{
\dontrun{
  PlotData <- Quantico:::PreparePlotData(input, TargetVariable = "TargetVariables", DateVariable = "DateVariables", GroupVariables = GroupVariables, G1Levels = "TS_Group1Levels", G2Levels = "TS_Group2Levels", G3Levels = "TS_Group3Levels")
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
\keyword{internal}
