% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Grow_Modal_Fun}
\alias{Grow_Modal_Fun}
\title{Grow_Modal_Fun}
\usage{
Grow_Modal_Fun(
  id,
  AppWidth = 12L,
  JoinData_ByXVariables_Selected = NULL,
  JoinData_ByYVariables_Selected = NULL,
  JoinData_SelectData1_Choices = NULL,
  JoinData_SelectData2_Choices = NULL,
  JoinData_JoinType_Choices = NULL,
  UnionData_SelectData1_Choices = NULL,
  UnionData_SelectData2_Choices = NULL,
  UnionData_Fill_Choices = NULL,
  UnionData_UseNames_Choices = NULL,
  JoinData_NewName_Selected = NULL,
  JoinData_SelectData1_Selected = NULL,
  JoinData_SelectData2_Selected = NULL,
  JoinData_JoinType_Selected = NULL,
  UnionData_NewName_Selected = NULL,
  UnionData_SelectData1_Selected = NULL,
  UnionData_SelectData2_Selected = NULL
)
}
\arguments{
\item{id}{= 'Grow_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Join and Union data
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
