% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{H2O_DRF_Modal_Fun}
\alias{H2O_DRF_Modal_Fun}
\title{H2O_DRF_Modal_Fun}
\usage{
H2O_DRF_Modal_Fun(
  id,
  AppWidth = 12L,
  H2O_DRF_eval_metric_Selected = NULL,
  H2O_DRF_TargetColumnName_Selected = NULL,
  H2O_DRF_FeatureColNames_Selected = NULL,
  H2O_DRF_TransformNumericColumns_Selected = NULL,
  H2O_DRF_WeightsColumn_Selected = NULL,
  H2O_DRF_Runs_Choices = NULL,
  H2O_DRF_TargetType_Choices = NULL,
  H2O_DRF_data_Choices = NULL,
  H2O_DRF_ValidationData_Choices = NULL,
  H2O_DRF_TestData_Choices = NULL,
  H2O_DRF_EncodeMethod_Choices = NULL,
  H2O_DRF_Methods_Choices = NULL,
  H2O_DRF_NThreads_Choices = NULL,
  H2O_DRF_MaxMem_Choices = NULL,
  H2O_DRF_TrainOnFull_Choices = NULL,
  H2O_DRF_Trees_Choices = NULL,
  H2O_DRF_MaxDepth_Choices = NULL,
  H2O_DRF_SampleRate_Choices = NULL,
  H2O_DRF_MTries_Choices = NULL,
  H2O_DRF_ColSampleRatePerTree_Choices = NULL,
  H2O_DRF_MinRows_Choices = NULL,
  H2O_DRF_NBinsCats_Choices = NULL,
  H2O_DRF_NBinsTopLevel_Choices = NULL,
  H2O_DRF_HistogramType_Choices = NULL,
  H2O_DRF_CategoricalEncoding_Choices = NULL,
  H2O_DRF_StoppingRounds_Choices = NULL,
  H2O_DRF_GridTune_Choices = NULL,
  H2O_DRF_MaxModelsInGrid_Choices = NULL,
  H2O_DRF_ModelID_Selected = NULL,
  H2O_DRF_Runs_Selected = NULL,
  H2O_DRF_TargetType_Selected = NULL,
  H2O_DRF_data_Selected = NULL,
  H2O_DRF_ValidationData_Selected = NULL,
  H2O_DRF_TestData_Selected = NULL,
  H2O_DRF_EncodeMethod_Selected = NULL,
  H2O_DRF_Methods_Selected = NULL,
  H2O_DRF_NThreads_Selected = NULL,
  H2O_DRF_MaxMem_Selected = NULL,
  H2O_DRF_TrainOnFull_Selected = NULL,
  H2O_DRF_Trees_Selected = NULL,
  H2O_DRF_MaxDepth_Selected = NULL,
  H2O_DRF_SampleRate_Selected = NULL,
  H2O_DRF_MTries_Selected = NULL,
  H2O_DRF_ColSampleRatePerTree_Selected = NULL,
  H2O_DRF_MinRows_Selected = NULL,
  H2O_DRF_NBinsCats_Selected = NULL,
  H2O_DRF_NBinsTopLevel_Selected = NULL,
  H2O_DRF_HistogramType_Selected = NULL,
  H2O_DRF_CategoricalEncoding_Selected = NULL,
  H2O_DRF_StoppingRounds_Selected = NULL,
  H2O_DRF_GridTune_Selected = NULL,
  H2O_DRF_MaxModelsInGrid_Selected = NULL,
  H2O_DRF_ClassWeights0_Selected = NULL,
  H2O_DRF_ClassWeights1_Selected = NULL
)
}
\arguments{
\item{id}{= 'H2O_DRF_ID'}

\item{AppWidth}{= 12L}
}
\description{
H2O DRF Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
