% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{H2O_GLM_Modal_Fun}
\alias{H2O_GLM_Modal_Fun}
\title{H2O_GLM_Modal_Fun}
\usage{
H2O_GLM_Modal_Fun(
  id,
  AppWidth = 12L,
  H2O_GLM_TargetColumnName_Selected = NULL,
  H2O_GLM_FeatureColNames_Selected = NULL,
  H2O_GLM_TransformNumericColumns_Selected = NULL,
  H2O_GLM_WeightsColumn_Selected = NULL,
  H2O_GLM_Distribution_Selected = NULL,
  H2O_GLM_Link_Selected = NULL,
  H2O_GLM_eval_metric_Selected = NULL,
  H2O_GLM_Runs_Choices = NULL,
  H2O_GLM_TargetType_Choices = NULL,
  H2O_GLM_data_Choices = NULL,
  H2O_GLM_ValidationData_Choices = NULL,
  H2O_GLM_TestData_Choices = NULL,
  H2O_GLM_EncodeMethod_Choices = NULL,
  H2O_GLM_Methods_Choices = NULL,
  H2O_GLM_NThreads_Choices = NULL,
  H2O_GLM_MaxMem_Choices = NULL,
  H2O_GLM_TrainOnFull_Choices = NULL,
  H2O_GLM_Solver_Choices = NULL,
  H2O_GLM_Alpha_Choices = NULL,
  H2O_GLM_Lambda_Choices = NULL,
  H2O_GLM_LambdaSearch_Choices = NULL,
  H2O_GLM_NLambdas_Choices = NULL,
  H2O_GLM_Standardize_Choices = NULL,
  H2O_GLM_RemoveCollinearColumns_Choices = NULL,
  H2O_GLM_InterceptInclude_Choices = NULL,
  H2O_GLM_NonNegativeCoefficients_Choices = NULL,
  H2O_GLM_TweedieLinkPower_Choices = NULL,
  H2O_GLM_TweedieVariancePower_Choices = NULL,
  H2O_GLM_GridTune_Choices = NULL,
  H2O_GLM_MaxModelsInGrid_Choices = NULL,
  H2O_GLM_ModelID_Selected = NULL,
  H2O_GLM_Runs_Selected = NULL,
  H2O_GLM_TargetType_Selected = NULL,
  H2O_GLM_data_Selected = NULL,
  H2O_GLM_ValidationData_Selected = NULL,
  H2O_GLM_TestData_Selected = NULL,
  H2O_GLM_EncodeMethod_Selected = NULL,
  H2O_GLM_Methods_Selected = NULL,
  H2O_GLM_NThreads_Selected = NULL,
  H2O_GLM_MaxMem_Selected = NULL,
  H2O_GLM_TrainOnFull_Selected = NULL,
  H2O_GLM_Solver_Selected = NULL,
  H2O_GLM_Alpha_Selected = NULL,
  H2O_GLM_Lambda_Selected = NULL,
  H2O_GLM_LambdaSearch_Selected = NULL,
  H2O_GLM_NLambdas_Selected = NULL,
  H2O_GLM_Standardize_Selected = NULL,
  H2O_GLM_RemoveCollinearColumns_Selected = NULL,
  H2O_GLM_InterceptInclude_Selected = NULL,
  H2O_GLM_NonNegativeCoefficients_Selected = NULL,
  H2O_GLM_TweedieLinkPower_Selected = NULL,
  H2O_GLM_TweedieVariancePower_Selected = NULL,
  H2O_GLM_GridTune_Selected = NULL,
  H2O_GLM_MaxModelsInGrid_Selected = NULL,
  H2O_GLM_ClassWeights0_Selected = NULL,
  H2O_GLM_ClassWeights1_Selected = NULL
)
}
\arguments{
\item{id}{= 'H2O_GLM_ID'}

\item{AppWidth}{= 12L}
}
\description{
H2O GLM Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
