% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{H2O_HGLM_Modal_Fun}
\alias{H2O_HGLM_Modal_Fun}
\title{H2O_HGLM_Modal_Fun}
\usage{
H2O_HGLM_Modal_Fun(
  id,
  AppWidth = 12L,
  H2O_HGLM_TargetColumnName_Selected = NULL,
  H2O_HGLM_FeatureColNames_Selected = NULL,
  H2O_HGLM_RandomColNumbers_Selected = NULL,
  H2O_HGLM_TransformNumericColumns_Selected = NULL,
  H2O_HGLM_WeightsColumn_Selected = NULL,
  H2O_HGLM_Distribution_Selected = NULL,
  H2O_HGLM_Link_Selected = NULL,
  H2O_HGLM_eval_metric_Selected = NULL,
  H2O_HGLM_Runs_Choices = NULL,
  H2O_HGLM_TargetType_Choices = NULL,
  H2O_HGLM_data_Choices = NULL,
  H2O_HGLM_ValidationData_Choices = NULL,
  H2O_HGLM_TestData_Choices = NULL,
  H2O_HGLM_EncodeMethod_Choices = NULL,
  H2O_HGLM_Methods_Choices = NULL,
  H2O_HGLM_NThreads_Choices = NULL,
  H2O_HGLM_MaxMem_Choices = NULL,
  H2O_HGLM_TrainOnFull_Choices = NULL,
  H2O_HGLM_Solver_Choices = NULL,
  H2O_HGLM_Alpha_Choices = NULL,
  H2O_HGLM_Lambda_Choices = NULL,
  H2O_HGLM_LambdaSearch_Choices = NULL,
  H2O_HGLM_NLambdas_Choices = NULL,
  H2O_HGLM_Standardize_Choices = NULL,
  H2O_HGLM_RemoveCollinearColumns_Choices = NULL,
  H2O_HGLM_InterceptInclude_Choices = NULL,
  H2O_HGLM_NonNegativeCoefficients_Choices = NULL,
  H2O_HGLM_RandomDistribution_Choices = NULL,
  H2O_HGLM_RandomLink_Choices = NULL,
  H2O_HGLM_TweedieLinkPower_Choices = NULL,
  H2O_HGLM_TweedieVariancePower_Choices = NULL,
  H2O_HGLM_GridTune_Choices = NULL,
  H2O_HGLM_MaxModelsInGrid_Choices = NULL,
  H2O_HGLM_ModelID_Selected = NULL,
  H2O_HGLM_Runs_Selected = NULL,
  H2O_HGLM_TargetType_Selected = NULL,
  H2O_HGLM_data_Selected = NULL,
  H2O_HGLM_ValidationData_Selected = NULL,
  H2O_HGLM_TestData_Selected = NULL,
  H2O_HGLM_EncodeMethod_Selected = NULL,
  H2O_HGLM_Methods_Selected = NULL,
  H2O_HGLM_NThreads_Selected = NULL,
  H2O_HGLM_MaxMem_Selected = NULL,
  H2O_HGLM_TrainOnFull_Selected = NULL,
  H2O_HGLM_Solver_Selected = NULL,
  H2O_HGLM_Alpha_Selected = NULL,
  H2O_HGLM_Lambda_Selected = NULL,
  H2O_HGLM_LambdaSearch_Selected = NULL,
  H2O_HGLM_NLambdas_Selected = NULL,
  H2O_HGLM_Standardize_Selected = NULL,
  H2O_HGLM_RemoveCollinearColumns_Selected = NULL,
  H2O_HGLM_InterceptInclude_Selected = NULL,
  H2O_HGLM_NonNegativeCoefficients_Selected = NULL,
  H2O_HGLM_RandomDistribution_Selected = NULL,
  H2O_HGLM_RandomLink_Selected = NULL,
  H2O_HGLM_TweedieLinkPower_Selected = NULL,
  H2O_HGLM_TweedieVariancePower_Selected = NULL,
  H2O_HGLM_GridTune_Selected = NULL,
  H2O_HGLM_MaxModelsInGrid_Selected = NULL,
  H2O_HGLM_ClassWeights0_Selected = NULL,
  H2O_HGLM_ClassWeights1_Selected = NULL
)
}
\arguments{
\item{id}{= 'H2O_HGLM_ID'}

\item{AppWidth}{= 12L}
}
\description{
H2O HGLM Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
