% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomePage.R
\name{HomePage}
\alias{HomePage}
\title{HomePage}
\usage{
HomePage(id, Page, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotPanels'}

\item{Page}{numeric}

\item{AppWidth}{= 12L}

\item{NumPlotsAvailable}{numeric}

\item{DragulaChoices}{character}
}
\description{
Panels
}
\author{
Adrian Antico
}
\concept{Plotting}
