% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{InferencePanels}
\alias{InferencePanels}
\title{InferencePanels}
\usage{
InferencePanels(id, Page, AppWidth = 12L, IOL = NULL)
}
\arguments{
\item{id}{= 'InferencePanels'}

\item{Page}{numeric}

\item{AppWidth}{= 12L}

\item{DL}{numeric}
}
\description{
Panels
}
\seealso{
Other Output Panels: 
\code{\link{EDAPanels}()},
\code{\link{MLPanels}()},
\code{\link{PlotDropDownContents}()},
\code{\link{PlotPanels}()}
}
\author{
Adrian Antico
}
\concept{Output Panels}
