% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{Install}
\alias{Install}
\title{Install}
\usage{
Install(Root = NULL)
}
\arguments{
\item{Root}{NULL will setwd to project root as defined in function}
}
\description{
To install the package
}
\seealso{
Other Utilities: 
\code{\link{BuildBinary}()},
\code{\link{UpdateDocs}()}
}
\author{
Adrian Antico
}
\concept{Utilities}
