% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUtils.R
\name{IntraSessionDefaults}
\alias{IntraSessionDefaults}
\title{IntraSessionDefaults}
\usage{
IntraSessionDefaults(
  List = NULL,
  InputName = "Plot1_SelectData",
  ArgName = "SelectedDefault",
  Default = NULL,
  Debug = Debug
)
}
\arguments{
\item{List}{named list of args where the args can be many}

\item{InputName}{'Plot1_SelectData' in input$Plot1_SelectData}

\item{ArgName}{'SelectedDefault' in List[[InputName]][[Default]]}

\item{Default}{default value to assign}

\item{Debug}{Logical}
}
\description{
Default generator
}
\seealso{
Other Utils: 
\code{\link{CleanVector}()},
\code{\link{DataTable}()},
\code{\link{Shiny.CodePrint.Collect}()},
\code{\link{Shiny.CodePrint.OrganizeCode}()},
\code{\link{Shiny.Utils.CachePath}()},
\code{\link{Utils.Code.Lines}()},
\code{\link{Utils.LinuxRunSH}()},
\code{\link{Utils.Nest}()}
}
\author{
Adrian Antico
}
\concept{Utils}
