% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Jarque.Bera.Test}
\alias{Jarque.Bera.Test}
\title{Jarque.Bera.Test}
\usage{
Jarque.Bera.Test(Vals, SampleSize = 5000, Samples = 30)
}
\arguments{
\item{Vals}{a vector of values to test}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Jarque.Bera.Test
}
\examples{
\dontrun{
output <- Jarque.Bera.Test(Vals = runif(100000), SampleSize = 1000, Samples = 500)
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
