% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Kolmogorov.Smirnov.Test}
\alias{Kolmogorov.Smirnov.Test}
\title{Kolmogorov.Smirnov.Test}
\usage{
Kolmogorov.Smirnov.Test(Vals, SampleSize = 500, Samples = 30)
}
\arguments{
\item{Vals}{a vector of values to test}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Kolmogorov.Smirnov.Test
}
\examples{
\dontrun{
output <- Kolmogorov.Smirnov.Test(runif(10000))
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
