% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{LevelsHex}
\alias{LevelsHex}
\title{LevelsHex}
\usage{
LevelsHex(dt, gv, name = "Blues")
}
\description{
Returns a vector of N color hex values with transparency added, where N is the number of unique levels in gv. Uses RColorBrewer to return base colors and then they are modified with transparency.
}
\examples{
\dontrun{
DataMuse:::LevelsHex(dt,gv,'Blues')
}
}
\keyword{internal}
