% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{LightGBM_Modal_Fun}
\alias{LightGBM_Modal_Fun}
\title{LightGBM_Modal_Fun}
\usage{
LightGBM_Modal_Fun(
  id,
  AppWidth = 12L,
  LightGBM_TargetColumnName_Selected = NULL,
  LightGBM_FeatureColNames_Selected = NULL,
  LightGBM_WeightsColumnName_Selected = NULL,
  LightGBM_TransformNumericColumns_Selected = NULL,
  LightGBM_grid_eval_metric_Selected = NULL,
  LightGBM_metric_Selected = NULL,
  LightGBM_Runs_Choices = NULL,
  LightGBM_TargetType_Choices = NULL,
  LightGBM_data_Choices = NULL,
  LightGBM_ValidationData_Choices = NULL,
  LightGBM_TestData_Choices = NULL,
  LightGBM_EncodeMethod_Choices = NULL,
  LightGBM_Methods_Choices = NULL,
  LightGBM_NThreads_Choices = NULL,
  LightGBM_device_type_Choices = NULL,
  LightGBM_NumGPUs_Choices = NULL,
  LightGBM_TrainOnFull_Choices = NULL,
  LightGBM_ModelID_Choices = NULL,
  LightGBM_Trees_Choices = NULL,
  LightGBM_max_depth_Choices = NULL,
  LightGBM_eta_Choices = NULL,
  LightGBM_min_data_in_leaf_Choices = NULL,
  LightGBM_num_leaves_Choices = NULL,
  LightGBM_bagging_fraction_Choices = NULL,
  LightGBM_feature_fraction_Choices = NULL,
  LightGBM_feature_fraction_bynode_Choices = NULL,
  LightGBM_lambda_l1_Choices = NULL,
  LightGBM_lambda_l2_Choices = NULL,
  LightGBM_PassInGrid_Choices = NULL,
  LightGBM_GridTune_Choices = NULL,
  LightGBM_MaxModelsInGrid_Choices = NULL,
  LightGBM_MaxRunsWithoutNewWinner_Choices = NULL,
  LightGBM_MaxRunMinutes_Choices = NULL,
  LightGBM_BaselineComparison_Choices = NULL,
  LightGBM_ModelID_Selected = NULL,
  LightGBM_Runs_Selected = NULL,
  LightGBM_TargetType_Selected = NULL,
  LightGBM_data_Selected = NULL,
  LightGBM_ValidationData_Selected = NULL,
  LightGBM_TestData_Selected = NULL,
  LightGBM_EncodeMethod_Selected = NULL,
  LightGBM_Methods_Selected = NULL,
  LightGBM_NThreads_Selected = NULL,
  LightGBM_device_type_Selected = NULL,
  LightGBM_NumGPUs_Selected = NULL,
  LightGBM_TrainOnFull_Selected = NULL,
  LightGBM_Trees_Selected = NULL,
  LightGBM_max_depth_Selected = NULL,
  LightGBM_eta_Selected = NULL,
  LightGBM_min_data_in_leaf_Selected = NULL,
  LightGBM_num_leaves_Selected = NULL,
  LightGBM_bagging_fraction_Selected = NULL,
  LightGBM_feature_fraction_Selected = NULL,
  LightGBM_feature_fraction_bynode_Selected = NULL,
  LightGBM_lambda_l1_Selected = NULL,
  LightGBM_lambda_l2_Selected = NULL,
  LightGBM_PassInGrid_Selected = NULL,
  LightGBM_GridTune_Selected = NULL,
  LightGBM_MaxModelsInGrid_Selected = NULL,
  LightGBM_MaxRunsWithoutNewWinner_Selected = NULL,
  LightGBM_MaxRunMinutes_Selected = NULL,
  LightGBM_BaselineComparison_Selected = NULL,
  LightGBM_ClassWeights0_Selected = NULL,
  LightGBM_ClassWeights1_Selected = NULL
)
}
\arguments{
\item{id}{= 'LightGBM_ID'}

\item{AppWidth}{= 12L}
}
\description{
LightGBM Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
