% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{LightGBMCARMA_Modal_Fun}
\alias{LightGBMCARMA_Modal_Fun}
\title{LightGBMCARMA_Modal_Fun}
\usage{
LightGBMCARMA_Modal_Fun(
  id,
  AppWidth = 12L,
  LightGBMCARMA_ModelID_Selected = NULL,
  LightGBMCARMA_TargetColumnName_Selected = NULL,
  LightGBMCARMA_DateColumnName_Selected = NULL,
  LightGBMCARMA_GroupVariables_Selected = NULL,
  LightGBMCARMA_RunMode_Choices = NULL,
  LightGBMCARMA_TimeUnit_Choices = NULL,
  LightGBMCARMA_FCPeriods_Choices = NULL,
  LightGBMCARMA_TaskType_Choices = NULL,
  LightGBMCARMA_NumGPU_Choices = NULL,
  LightGBMCARMA_SaveModel_Choices = NULL,
  LightGBMCARMA_ArgsList_Choices = NULL,
  LightGBMCARMA_TimeWeights_Choices = NULL,
  LightGBMCARMA_Methods_Choices = NULL,
  LightGBMCARMA_Difference_Choices = NULL,
  LightGBMCARMA_NonNegativePred_Choices = NULL,
  LightGBMCARMA_RoundPreds_Choices = NULL,
  LightGBMCARMA_CalendarVariables_Choices = NULL,
  LightGBMCARMA_HolidayVariables_Choices = NULL,
  LightGBMCARMA_Lookback_Choices = NULL,
  LightGBMCARMA_Lags_Choices = NULL,
  LightGBMCARMA_MovingAverages_Choices = NULL,
  LightGBMCARMA_MovingSD_Choices = NULL,
  LightGBMCARMA_MovingSkew_Choices = NULL,
  LightGBMCARMA_MovingKurt_Choices = NULL,
  LightGBMCARMA_MovingQuantiles_Choices = NULL,
  LightGBMCARMA_Quantiles_Selected_Choices = NULL,
  LightGBMCARMA_AnomalyDetection_HighThreshold_Choices = NULL,
  LightGBMCARMA_AnomalyDetection_LowThreshold_Choices = NULL,
  LightGBMCARMA_TimeTrend_Choices = NULL,
  LightGBMCARMA_DataTruncate_Choices = NULL,
  LightGBMCARMA_EncodingMethod_Choices = NULL,
  LightGBMCARMA_ZeroPadSeries_Choices = NULL,
  LightGBMCARMA_CrossEvalModelUpdate_Choices = NULL,
  LightGBMCARMA_CrossEvalValuesUpdate_Choices = NULL,
  LightGBMCARMA_CrossEvalModelUpdateFreq_Choices = NULL,
  LightGBMCARMA_CrossEvalValuesUpdateFreq_Choices = NULL,
  LightGBMCARMA_EvalMetric_Choices = NULL,
  LightGBMCARMA_LossFunction_Choices = NULL,
  LightGBMCARMA_TrainData_Choices = NULL,
  LightGBMCARMA_ValidationData_Choices = NULL,
  LightGBMCARMA_XREGS_Choices = NULL,
  LightGBMCARMA_Trees_Choices = NULL,
  LightGBMCARMA_Max_Depth_Choices = NULL,
  LightGBMCARMA_ETA_Choices = NULL,
  LightGBMCARMA_Min_Data_In_Leaf_Choices = NULL,
  LightGBMCARMA_Num_Leaves_Choices = NULL,
  LightGBMCARMA_Bagging_Fraction_Choices = NULL,
  LightGBMCARMA_Feature_Fraction_Choices = NULL,
  LightGBMCARMA_Feature_Fraction_Bynode_Choices = NULL,
  LightGBMCARMA_Lambda_L1_Choices = NULL,
  LightGBMCARMA_Lambda_L2_Choices = NULL,
  LightGBMCARMA_RunMode_Selected = NULL,
  LightGBMCARMA_TimeUnit_Selected = NULL,
  LightGBMCARMA_FCPeriods_Selected = NULL,
  LightGBMCARMA_TaskType_Selected = NULL,
  LightGBMCARMA_NumGPU_Selected = NULL,
  LightGBMCARMA_SaveModel_Selected = NULL,
  LightGBMCARMA_ArgsList_Selected = NULL,
  LightGBMCARMA_TimeWeights_Selected = NULL,
  LightGBMCARMA_Methods_Selected = NULL,
  LightGBMCARMA_Difference_Selected = NULL,
  LightGBMCARMA_NonNegativePred_Selected = NULL,
  LightGBMCARMA_RoundPreds_Selected = NULL,
  LightGBMCARMA_CalendarVariables_Selected = NULL,
  LightGBMCARMA_HolidayVariables_Selected = NULL,
  LightGBMCARMA_Lookback_Selected = NULL,
  LightGBMCARMA_Lags_Selected = NULL,
  LightGBMCARMA_MovingAverages_Selected = NULL,
  LightGBMCARMA_MovingSD_Selected = NULL,
  LightGBMCARMA_MovingSkew_Selected = NULL,
  LightGBMCARMA_MovingKurt_Selected = NULL,
  LightGBMCARMA_MovingQuantiles_Selected = NULL,
  LightGBMCARMA_Quantiles_Selected_Selected = NULL,
  LightGBMCARMA_AnomalyDetection_HighThreshold_Selected = NULL,
  LightGBMCARMA_AnomalyDetection_LowThreshold_Selected = NULL,
  LightGBMCARMA_TimeTrend_Selected = NULL,
  LightGBMCARMA_DataTruncate_Selected = NULL,
  LightGBMCARMA_EncodingMethod_Selected = NULL,
  LightGBMCARMA_ZeroPadSeries_Selected = NULL,
  LightGBMCARMA_CrossEvalModelUpdate_Selected = NULL,
  LightGBMCARMA_CrossEvalValuesUpdate_Selected = NULL,
  LightGBMCARMA_CrossEvalModelUpdateFreq_Selected = NULL,
  LightGBMCARMA_CrossEvalValuesUpdateFreq_Selected = NULL,
  LightGBMCARMA_EvalMetric_Selected = NULL,
  LightGBMCARMA_LossFunction_Selected = NULL,
  LightGBMCARMA_TrainData_Selected = NULL,
  LightGBMCARMA_ValidationData_Selected = NULL,
  LightGBMCARMA_XREGS_Selected = NULL,
  LightGBMCARMA_Trees_Selected = NULL,
  LightGBMCARMA_Max_Depth_Selected = NULL,
  LightGBMCARMA_ETA_Selected = NULL,
  LightGBMCARMA_Min_Data_In_Leaf_Selected = NULL,
  LightGBMCARMA_Num_Leaves_Selected = NULL,
  LightGBMCARMA_Bagging_Fraction_Selected = NULL,
  LightGBMCARMA_Feature_Fraction_Selected = NULL,
  LightGBMCARMA_Feature_Fraction_Bynode_Selected = NULL,
  LightGBMCARMA_Lambda_L1_Selected = NULL,
  LightGBMCARMA_Lambda_L2_Selected = NULL
)
}
\arguments{
\item{id}{= 'LightGBMCARMA_ID'}

\item{AppWidth}{= 12L}
}
\description{
LightGBM CARMA Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoostCARMA_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
