% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{LMER}
\alias{LMER}
\title{LMER}
\usage{
LMER(
  data,
  TargetVariable = NULL,
  FixedEffects = NULL,
  RandomEffects = NULL,
  Type = "add",
  Poly = NULL,
  Family = "gaussian"
)
}
\arguments{
\item{data}{data.table}

\item{TargetVariable}{target variable name}

\item{FixedEffects}{NULL. Optional}

\item{RandomEffects}{NULL. Must include at least one variable}

\item{Poly}{NULL. If set to a numeric value, then polynomial terms will be used for the Fixed Effects. Used for when the FixedEffect is a Prediction output and a single value.}

\item{Family}{= 'gaussian' or 'binomial'}
}
\description{
LM with coefficients greater than or equal to zero if Constrained is TRUE. Otherwise, a fast lm coefficient set is returned from armadillo
}
\details{
Up to 5 random effects variables can be included.

There are
}
\examples{
FixedEffects <- 'Predict'; RandomEffects <- 'Factor_1', TargetVariable <- 'TargetVar'

}
\seealso{
Other Mixed Effects: 
\code{\link{EffectsInterval}()},
\code{\link{FastLM}()},
\code{\link{MEOW}()},
\code{\link{PredictionIntervals}()},
\code{\link{REsim}()}
}
\author{
Adrian Antico
}
\concept{Mixed Effects}
