% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Local_Modal_Fun}
\alias{Local_Modal_Fun}
\title{Local_Modal_Fun}
\usage{
Local_Modal_Fun(
  id,
  AppWidth = 12L,
  SaveData_SelectData_Choices = NULL,
  ExportFileType_Choices = NULL,
  SaveData_SelectDataRdata_Choices = NULL,
  SaveData_SelectData_Selected = NULL,
  SaveFileName_Selected = NULL,
  ExportFileType_Selected = NULL,
  SaveData_SelectDataRdata_Selected = NULL,
  SaveFileNameRdata_Selected = NULL
)
}
\arguments{
\item{id}{= 'LocalDMID'}

\item{AppWidth}{= 12L}
}
\description{
Local Environment
}
\seealso{
Other DM: 
\code{\link{AzureBlob_Modal_Fun}()},
\code{\link{DM_AzureBlob}()},
\code{\link{DM_Local}()},
\code{\link{DM_PostGRE}()},
\code{\link{DM_Session}()},
\code{\link{PostGRE_Modal_Fun}()},
\code{\link{Session_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DM}
