% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{LoginButton}
\alias{LoginButton}
\title{LoginButton}
\usage{
LoginButton(
  id = "LoginButton",
  Label = "Check Credentials",
  width = 1L,
  Icon = shiny::icon("chevron-right", lib = "font-awesome"),
  Style = "gradient",
  Color = "royal"
)
}
\arguments{
\item{id}{'Check_Credentials'}

\item{Label}{= 'Check Credentials',}

\item{width}{= 1L}

\item{Icon}{= shiny::icon('chevron-right', lib = 'font-awesome')}

\item{Style}{= 'gradient'}

\item{Color}{= 'royal'}
}
\description{
Action Button for Checking Credentials
}
\seealso{
Other Login: 
\code{\link{LoginInputs}()},
\code{\link{LogoutButton}()}
}
\author{
Adrian Antico
}
\concept{Login}
