% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{LoginInputs}
\alias{LoginInputs}
\title{LoginInputs}
\usage{
LoginInputs(
  id = "LoginBox",
  BoxStatus = "danger",
  UserNameLabel = "Select from Names",
  PassWordLabel = "Input Password",
  UserNameChoices = Credentials[["UserName"]],
  BoxTitle = NULL,
  SolidHeader = TRUE,
  Collapsible = FALSE
)
}
\arguments{
\item{id}{inputId}

\item{BoxStatus}{= 'danger'}

\item{UserNameLabel}{= "UserName"}

\item{PassWordLabel}{= 'Input Password'}

\item{UserNameChoices}{= Credentials[['UserName']]}

\item{BoxTitle}{= NULL}

\item{SolidHeader}{= TRUE}

\item{Collapsible}{= FALSE}
}
\description{
Box with UserName and PassWord Inputs. Username inputId is 'UserName'. PassWord inputId is 'Password'
}
\seealso{
Other Login: 
\code{\link{LoginButton}()},
\code{\link{LogoutButton}()}
}
\author{
Adrian Antico
}
\concept{Login}
