% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{LogoutButton}
\alias{LogoutButton}
\title{LogoutButton}
\usage{
LogoutButton(
  id = "LogoutButton",
  Label = "Reset Session",
  width = 1L,
  Icon = NULL,
  Style = "gradient",
  Color = "royal"
)
}
\arguments{
\item{id}{'Check_Credentials'}

\item{Label}{= 'Check Credentials',}

\item{width}{= 1L}

\item{Icon}{= shiny::icon('chevron-right', lib = 'font-awesome')}

\item{Style}{= 'gradient'}

\item{Color}{= 'royal'}
}
\description{
Action Button for Checking Credentials
}
\seealso{
Other Login: 
\code{\link{LoginButton}()},
\code{\link{LoginInputs}()}
}
\author{
Adrian Antico
}
\concept{Login}
