% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Misc_Modal_Fun}
\alias{Misc_Modal_Fun}
\title{Misc_Modal_Fun}
\usage{
Misc_Modal_Fun(
  id,
  AppWidth = 12L,
  TSF_DateColumnName_Selected = NULL,
  TSF_GroupVariables_Selected = NULL,
  Roll_RollVars_Selected = NULL,
  Roll_NonRollVars_Selected = NULL,
  Roll_DateColumnName_Selected = NULL,
  Roll_GroupVariables_Selected = NULL,
  MetaProgramming_TextCode_Choices = NULL,
  TSF_SelectData_Choices = NULL,
  TSF_TimeUnit_Choices = NULL,
  TSF_MaxMissingPercent_Choices = NULL,
  TSF_SimpleImpute_Choices = NULL,
  Roll_SelectData_Choices = NULL,
  Roll_TimeUnit_Choices = NULL,
  Roll_RollDirection_Choices = NULL,
  Roll_SimpleImpute_Choices = NULL,
  TSF_SelectData_Selected = NULL,
  TSF_NewName_Selected = NULL,
  TSF_TimeUnit_Selected = NULL,
  TSF_MaxMissingPercent_Selected = NULL,
  TSF_SimpleImpute_Selected = NULL,
  Roll_SelectData_Selected = NULL,
  Roll_NewName_Selected = NULL,
  Roll_TimeUnit_Selected = NULL,
  Roll_RollDirection_Selected = NULL,
  Roll_SimpleImpute_Selected = NULL
)
}
\arguments{
\item{id}{= 'Misc_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Misc operations
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
