% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyActionButtons.R
\name{ML_CausalMediation}
\alias{ML_CausalMediation}
\title{ML_CausalMediation}
\usage{
ML_CausalMediation(
  id = "ML_CausalMediationML",
  Style = "gradient",
  Color = "royal",
  Align = "center",
  DropDownRight = TRUE,
  Animate = TRUE,
  Status = "custom",
  H3Color = "white"
)
}
\arguments{
\item{id}{= 'ML_CausalMediation'}

\item{Style}{= 'gradient',}

\item{Color}{= 'royal',}

\item{Align}{= 'center'}

\item{DropDownRight}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}

\item{H3Color}{= 'white'}
}
\description{
CatBoost FC button
}
\seealso{
Other Modeling: 
\code{\link{BuildModelsButton}()},
\code{\link{FC_Arfima}()},
\code{\link{FC_CatBoostCARMAFUN}()},
\code{\link{FC_CatBoostCARMAID}()},
\code{\link{FC_CatBoostCARMA}()},
\code{\link{FC_ETS}()},
\code{\link{FC_LightGBMCARMAFUN}()},
\code{\link{FC_LightGBMCARMAID}()},
\code{\link{FC_LightGBMCARMA}()},
\code{\link{FC_NNet}()},
\code{\link{FC_SARIMA}()},
\code{\link{FC_TBATS}()},
\code{\link{FC_XGBoostCARMAFUN}()},
\code{\link{FC_XGBoostCARMAID}()},
\code{\link{FC_XGBoostCARMA}()},
\code{\link{ML_CatBoostReport}()},
\code{\link{ML_CatBoost}()},
\code{\link{ML_GEE}()},
\code{\link{ML_H2O_DRF}()},
\code{\link{ML_H2O_GBM}()},
\code{\link{ML_H2O_GLM}()},
\code{\link{ML_H2O_HGLM}()},
\code{\link{ML_LightGBMReport}()},
\code{\link{ML_LightGBM}()},
\code{\link{ML_MixedEffects}()},
\code{\link{ML_Scoring}()},
\code{\link{ML_XGBoostReport}()},
\code{\link{ML_XGBoost}()}
}
\author{
Adrian Antico
}
\concept{Modeling}
