% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{MLPanels}
\alias{MLPanels}
\title{MLPanels}
\usage{
MLPanels(id, Page, AppWidth = 12L, MOL = NULL)
}
\arguments{
\item{id}{= 'MLPanels'}

\item{Page}{numeric}

\item{AppWidth}{= 12L}

\item{DL}{numeric}
}
\description{
Panels
}
\seealso{
Other Output Panels: 
\code{\link{EDAPanels}()},
\code{\link{PlotDropDownContents}()},
\code{\link{PlotPanels}()}
}
\author{
Adrian Antico
}
\concept{Output Panels}
