% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{NLP_Modal_Fun}
\alias{NLP_Modal_Fun}
\title{NLP_Modal_Fun}
\usage{
NLP_Modal_Fun(
  id,
  AppWidth = 12L,
  Word2Vec_H2O_stringCol_Selected = NULL,
  TextSummary_TextColumns_Selected = NULL,
  Sentiment_TextColumns_Selected = NULL,
  Sentiment_CombineTextGroupVar_Selected = NULL,
  Readability_TextColumns_Selected = NULL,
  LexicalDiversity_TextColumns_Selected = NULL,
  Word2Vec_H2O_TrainData_Choices = NULL,
  Word2Vec_H2O_ValidationData_Choices = NULL,
  Word2Vec_H2O_TestData_Choices = NULL,
  Word2Vec_H2O_BuildType_Choices = NULL,
  Word2Vec_H2O_KeepStringCol_Choices = NULL,
  Word2Vec_H2O_vects_Choices = NULL,
  Word2Vec_H2O_MinWords_Choices = NULL,
  Word2Vec_H2O_WindowSize_Choices = NULL,
  Word2Vec_H2O_Epochs_Choices = NULL,
  TextSummary_RemoveStats_Choices = NULL,
  TextSummary_SelectData_Choices = NULL,
  Sentiment_SelectData_Choices = NULL,
  Sentiment_Response_Choices = NULL,
  Sentiment_RemoveStopWords_Choices = NULL,
  Sentiment_Stemming_Choices = NULL,
  Readability_SelectData_Choices = NULL,
  Readability_Measures_Choices = NULL,
  Readability_RemoveHyphens_Choices = NULL,
  Readability_MinSentenceLength_Choices = NULL,
  Readability_MaxSentenceLength_Choices = NULL,
  Readability_Intermediate_Choices = NULL,
  LexicalDiversity_SelectData_Choices = NULL,
  LexicalDiversity_Measures_Choices = NULL,
  LexicalDiversity_RemoveSymbols_Choices = NULL,
  LexicalDiversity_RemoveHyphens_Choices = NULL,
  LexicalDiversity_RemovePunctuation_Choices = NULL,
  LexicalDiversity_RemoveNumbers_Choices = NULL,
  LexicalDiversity_LogBase_Choices = NULL,
  LexicalDiversity_MATTR_Window_Choices = NULL,
  LexicalDiversity_MSTTR_Segment_Choices = NULL,
  Word2Vec_H2O_TrainData_Selected = NULL,
  Word2Vec_H2O_ValidationData_Selected = NULL,
  Word2Vec_H2O_TestData_Selected = NULL,
  Word2Vec_H2O_BuildType_Selected = NULL,
  Word2Vec_H2O_KeepStringCol_Selected = NULL,
  Word2Vec_H2O_vects_Selected = NULL,
  Word2Vec_H2O_MinWords_Selected = NULL,
  Word2Vec_H2O_WindowSize_Selected = NULL,
  Word2Vec_H2O_Epochs_Selected = NULL,
  TextSummary_RemoveStats_Selected = NULL,
  TextSummary_SelectData_Selected = NULL,
  Sentiment_SelectData_Selected = NULL,
  Sentiment_Response_Selected = NULL,
  Sentiment_RemoveStopWords_Selected = NULL,
  Sentiment_Stemming_Selected = NULL,
  Readability_SelectData_Selected = NULL,
  Readability_Measures_Selected = NULL,
  Readability_RemoveHyphens_Selected = NULL,
  Readability_MinSentenceLength_Selected = NULL,
  Readability_MaxSentenceLength_Selected = NULL,
  Readability_Intermediate_Selected = NULL,
  LexicalDiversity_SelectData_Selected = NULL,
  LexicalDiversity_Measures_Selected = NULL,
  LexicalDiversity_RemoveSymbols_Selected = NULL,
  LexicalDiversity_RemoveHyphens_Selected = NULL,
  LexicalDiversity_RemovePunctuation_Selected = NULL,
  LexicalDiversity_RemoveNumbers_Selected = NULL,
  LexicalDiversity_LogBase_Selected = NULL,
  LexicalDiversity_MATTR_Window_Selected = NULL,
  LexicalDiversity_MSTTR_Segment_Selected = NULL
)
}
\arguments{
\item{id}{= 'NLP_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
NLP methods
}
\seealso{
Other FE: 
\code{\link{AnomalyDetection_Modal_Fun}()},
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Categorical_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFillRoll}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_Clustering}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{Numeric_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
