% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Normality_Modal_Fun}
\alias{Normality_Modal_Fun}
\title{Normality_Modal_Fun}
\usage{
Normality_Modal_Fun(
  id,
  AppWidth = 12L,
  Normality_YVars_Selected = NULL,
  Normality_SelectData_Choices = NULL,
  Normality_SelectData_Selected = NULL,
  Normality_InferenceID_Selected = NULL,
  SampleSize.ADT_Selected = NULL,
  Samples.ADT_Selected = NULL,
  SampleSize.CVMT_Selected = NULL,
  Samples.CVMT_Selected = NULL,
  SampleSize.KST_Selected = NULL,
  Samples.KST_Selected = NULL,
  SampleSize.ST_Selected = NULL,
  Samples.ST_Selected = NULL,
  SampleSize.JBT_Selected = NULL,
  Samples.JBT_Selected = NULL,
  SampleSize.AT_Selected = NULL,
  Samples.AT_Selected = NULL
)
}
\arguments{
\item{id}{= 'Normality_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Inference operations
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Shapiro.Test}()}
}
\author{
Adrian Antico
}
\concept{Inference}
