% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Numeric_Modal_Fun}
\alias{Numeric_Modal_Fun}
\title{Numeric_Modal_Fun}
\usage{
Numeric_Modal_Fun(
  id,
  AppWidth = 12L,
  PercentRank_ColNames_Selected = NULL,
  PercentRank_GroupVars_Selected = NULL,
  Standardize_ColNames_Selected = NULL,
  Standardize_GroupVars_Selected = NULL,
  AutoInteraction_NumericVars_Selected = NULL,
  AutoTransformationCreate_ColumnNames_Selected = NULL,
  PercentRank_SelectData_Choices = NULL,
  PercentRank_SelectValidationData_Choices = NULL,
  PercentRank_SelectTestData_Choices = NULL,
  PercentRank_Granularity_Choices = NULL,
  Standardize_SelectData_Choices = NULL,
  Standardize_SelectValidationData_Choices = NULL,
  Standardize_SelectTestData_Choices = NULL,
  Standardize_Center_Choices = NULL,
  Standardize_Scale_Choices = NULL,
  AutoInteraction_SelectData_Choices = NULL,
  AutoInteraction_InteractionDepth_Choices = NULL,
  AutoInteraction_Scale_Choices = NULL,
  AutoInteraction_Center_Choices = NULL,
  AutoTransformationCreate_SelectData_Choices = NULL,
  AutoTransformationCreate_Methods_Choices = NULL,
  PercentRank_SelectData_Selected = NULL,
  PercentRank_SelectValidationData_Selected = NULL,
  PercentRank_SelectTestData_Selected = NULL,
  PercentRank_Granularity_Selected = NULL,
  Standardize_SelectData_Selected = NULL,
  Standardize_SelectValidationData_Selected = NULL,
  Standardize_SelectTestData_Selected = NULL,
  Standardize_Center_Selected = NULL,
  Standardize_Scale_Selected = NULL,
  AutoInteraction_SelectData_Selected = NULL,
  AutoInteraction_InteractionDepth_Selected = NULL,
  AutoInteraction_Scale_Selected = NULL,
  AutoInteraction_Center_Selected = NULL,
  AutoTransformationCreate_SelectData_Selected = NULL,
  AutoTransformationCreate_Methods_Selected = NULL
)
}
\arguments{
\item{id}{= 'Numeric_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Numeric variables
}
\seealso{
Other FE: 
\code{\link{AnomalyDetection_Modal_Fun}()},
\code{\link{Calendar_Modal_Fun}()},
\code{\link{Categorical_Modal_Fun}()},
\code{\link{Clustering_Modal_Fun}()},
\code{\link{DW_DataSets}()},
\code{\link{DimensionalityReduction_Modal_Fun}()},
\code{\link{FE.Impute.TimeSeriesFillRoll}()},
\code{\link{FE.Impute.TimeSeriesFill}()},
\code{\link{FE_AnomalyDetection}()},
\code{\link{FE_CategoricalVariables}()},
\code{\link{FE_Clustering}()},
\code{\link{FE_DateVariables}()},
\code{\link{FE_DimensionalityReduction}()},
\code{\link{FE_NLP}()},
\code{\link{FE_NumericVariables}()},
\code{\link{FE_Windowing}()},
\code{\link{NLP_Modal_Fun}()},
\code{\link{Shiny.FE.AnomalyDetection.IsolationForest.H2O}()},
\code{\link{Shiny.FE.Categorical.Dummify}()},
\code{\link{Shiny.FE.Clustering.Kmeans.H2O}()},
\code{\link{Shiny.FE.CrossRow.CategoricalEncoding}()},
\code{\link{Shiny.FE.CrossRow.Differencing}()},
\code{\link{Shiny.FE.CrossRow.RollingMode}()},
\code{\link{Shiny.FE.CrossRow.RollingStats}()},
\code{\link{Shiny.FE.Date.Calendar}()},
\code{\link{Shiny.FE.Date.Holiday}()},
\code{\link{Shiny.FE.DimReduction.AutoEncoder.H2O}()},
\code{\link{Shiny.FE.ModelDataPrep}()},
\code{\link{Shiny.FE.Numeric.Interactions}()},
\code{\link{Shiny.FE.Numeric.PercentRank}()},
\code{\link{Shiny.FE.Numeric.Standardize}()},
\code{\link{Shiny.FE.Numeric.Transformations}()},
\code{\link{Shiny.FE.PartitionData}()},
\code{\link{Shiny.FE.Word2Vec.H2O}()},
\code{\link{Shiny.NLP.LexicalDiversity}()},
\code{\link{Shiny.NLP.Readability}()},
\code{\link{Shiny.NLP.Sentiment}()},
\code{\link{Shiny.NLP.TextSummary}()},
\code{\link{Windowing_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{FE}
