% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyInputs.R
\name{NumericInput}
\alias{NumericInput}
\title{NumericInput}
\usage{
NumericInput(
  InputID = NULL,
  Update = FALSE,
  session = NULL,
  input = NULL,
  server = NULL,
  Label = NULL,
  PlotDropDown = NULL,
  Key = NULL,
  Step = 10,
  Value = 1,
  Min = 1,
  Max = 10
)
}
\arguments{
\item{InputID}{ID of the input}

\item{Update}{FALSE. Set to TRUE to run updateSelectizeInput}

\item{session}{For update usgae}

\item{input}{For update usage. Provide input if you want your inputs to update with previous values when screens reload}

\item{server}{For update usage}

\item{Label}{Feeds label}

\item{PlotDropDown}{From App}

\item{Key}{From App}

\item{Step}{Feeds size in the options list}

\item{Value}{Default}

\item{Min}{Min value}

\item{Max}{Max value}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
NumericInput automatically builds a numeric input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_CARMA_HolidayMovingAverages <- renderUI({
  DataMuse:::NumericInput(InputID = "TS_CARMA_HolidayMovingAverages",
                            Label = "Select Holiday Count MA's",
                            Min = 1,
                            Max = 1,
                            Value = 1,
                            Step = 1)})
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
