% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{One.Sample.TTest}
\alias{One.Sample.TTest}
\title{One.Sample.TTest}
\usage{
One.Sample.TTest(
  dt = NULL,
  YVar = NULL,
  NullValue = 0,
  Alternative = "two.sided",
  ConfidenceLevel = 0.95,
  SampleSize = NULL,
  Samples = 1,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "850px",
  PlotWidth = "1450px"
)
}
\arguments{
\item{dt}{data.table}

\item{YVar}{Variable to be tested}

\item{NullValue}{H0 value}

\item{Alternative}{"two.sided", "less", or "greater"}

\item{ConfidenceLevel}{numeric}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
One.Sample.TTest
}
\examples{
\dontrun{
dt = data.table::data.table(qnorm(p = runif(100000)))
YVar = "Daily Margin"
data.table::setnames(dt, "V1", YVar)
NullValue = 0
Alternative = "two.sided"
ConfidenceLevel = 0.95
SampleSize = 500
Samples = 50
EchartsTheme = "macarons"
TextColor = "black"
PlotHeight = "850px"
PlotWidth = "1450px"
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()}
}
\concept{Inference}
