% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{PanelFCReport}
\alias{PanelFCReport}
\title{PanelFCReport}
\usage{
PanelFCReport(DataList = NULL, MOL = NULL, ModelID = NULL, OutputPath = NULL)
}
\arguments{
\item{DataList}{DataList from app}

\item{MOL}{ModelObjectList from app}

\item{ModelID}{from app}

\item{OutputPath}{List of output objects}
}
\description{
PanelFCReport is an Rmarkdown report for viewing FC results
}
\seealso{
Other Reports: 
\code{\link{EDAReport}()},
\code{\link{ModelInsightsReport}()},
\code{\link{PlottingReport}()},
\code{\link{SSFCReport}()},
\code{\link{TablesReport}()}
}
\author{
Adrian Antico
}
\concept{Reports}
