% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Pearsons.Correlation.Test}
\alias{Pearsons.Correlation.Test}
\title{Pearsons.Correlation.Test}
\usage{
Pearsons.Correlation.Test(x, y, SampleSize, Samples)
}
\arguments{
\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}

\item{Vals}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}
}
\description{
Pearsons.Correlation.Test
}
\examples{
\dontrun{
x <- qnorm(p = runif(100000))
y <- qnorm(p = runif(100000))
Pearsons.Correlation.Test(x, y, SampleSize = 2500, Samples = 30)
}

}
\references{
https://en.m.wikipedia.org/wiki/Anderson%E2%80%93Darling_test, https://real-statistics.com/non-parametric-tests/goodness-of-fit-tests/anderson-darling-test/
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{All.Normality.Tests}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
