% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyInputs.R
\name{PickerInput}
\alias{PickerInput}
\title{PickerInput}
\usage{
PickerInput(
  session = NULL,
  input = NULL,
  Update = FALSE,
  InputID = NULL,
  Label = NULL,
  Choices = NULL,
  SelectedDefault = NULL,
  SubText = NULL,
  ContentList = NULL,
  Size = 18,
  SelectedText = "count > 10",
  PlotDropDown = NULL,
  Key = NULL,
  Multiple = TRUE,
  MaxVars = NULL,
  ActionBox = TRUE,
  PlotSelection = NULL,
  Debug = FALSE
)
}
\arguments{
\item{session}{For update usgae}

\item{input}{For update usage. Provide input if you want your inputs to update with previous values when screens reload}

\item{Update}{= FALSE}

\item{InputID}{ID of the input}

\item{Label}{Feeds label}

\item{Choices}{Feeds choices}

\item{SelectedDefault}{Feeds selected for cases where ProjectList has a null element}

\item{SubText}{= NULL}

\item{ContentList}{NULL}

\item{Size}{Feeds size in the options list}

\item{SelectedText}{Feeds selected-text-format in options list}

\item{PlotDropDown}{From App}

\item{Key}{From App}

\item{Multiple}{Feeds multiple for enabling selecting more than one element from list}

\item{ActionBox}{Feeds actions-box for option list}

\item{Debug}{FALSE}

\item{server}{For update usage}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
PickerInput automatically builds a picker input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_CARMA_HolidayMovingAverages <- renderUI({
  Quantico:::PickerInput(InputID = "TS_CARMA_HolidayMovingAverages", Label = "Select Holiday Count MA's", Choices = as.character(0:50),
                           SelectedDefault = as.character(c(1,2)), Size = 10, SelectedText = "count > 1", Multiple = TRUE, ActionBox = TRUE)})
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
