% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyInputs.R
\name{PickerInput_GetLevels2}
\alias{PickerInput_GetLevels2}
\title{PickerInput_GetLevels2}
\usage{
PickerInput_GetLevels2(
  session = NULL,
  input = NULL,
  Update = FALSE,
  InputID = NULL,
  PlotDropDown = NULL,
  Key = NULL,
  Choices = NULL,
  SelectedDefault = NULL,
  GroupVariable = NULL
)
}
\arguments{
\item{session}{= NULL}

\item{input}{= NULL}

\item{InputID}{Feeds ProjectList and inputId. Argument saved in ProjectList}

\item{PlotDropDown}{From App}

\item{Key}{From App}

\item{Choices}{Feeds choices}

\item{SelectedDefault}{Feeds selected for cases where ProjectList has a null element}

\item{GroupVariable}{From App}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
PickerInput_GetLevels2 automatically builds a picker input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_CARMA_HolidayMovingAverages <- renderUI({
  Quantico:::PickerInput_GetLevels2(
    InputID = "TS_CARMA_HolidayMovingAverages", Label = "Select Holiday Count MA's", Choices = as.character(0:50),
    SelectedDefault = as.character(c(1,2)), Multiple = TRUE)})
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
