% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Pivot_Modal_Fun}
\alias{Pivot_Modal_Fun}
\title{Pivot_Modal_Fun}
\usage{
Pivot_Modal_Fun(
  id,
  AppWidth = 12L,
  MeltData_id.vars_Selected = NULL,
  MeltData_measure.vars_Selected = NULL,
  CastData_id.vars_Selected = NULL,
  CastData_CastColumns_Selected = NULL,
  CastData_value.var_Selected = NULL,
  MeltData_SelectData_Choices = NULL,
  MeltData_na.rm_Choices = NULL,
  MeltData_variable.factor_Choices = NULL,
  MeltData_value.factor_Choices = NULL,
  CastData_SelectData_Choices = NULL,
  CastData_fun.aggregate_Choices = NULL,
  CastData_fill_Choices = NULL,
  MeltData_NewName_Selected = NULL,
  MeltData_SelectData_Selected = NULL,
  MeltData_na.rm_Selected = NULL,
  MeltData_variable.name_Selected = NULL,
  MeltData_value.name_Selected = NULL,
  MeltData_variable.factor_Selected = NULL,
  MeltData_value.factor_Selected = NULL,
  CastData_SelectData_Selected = NULL,
  CastData_NewName_Selected = NULL,
  CastData_fun.aggregate_Selected = NULL,
  CastData_fill_Selected = NULL
)
}
\arguments{
\item{id}{= 'Pivot_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Melt and cast data
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()},
\code{\link{Shrink_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DW}
