% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{PlotDropDownContents}
\alias{PlotDropDownContents}
\title{PlotDropDownContents}
\usage{
PlotDropDownContents(
  id,
  PlotNumber = 1,
  Align = "center",
  DropDownRight = FALSE,
  Animate = TRUE,
  Status = "custom",
  H3Color = "#0088a7"
)
}
\arguments{
\item{id}{= 'PlotDropDown'}

\item{PlotNumber}{= 1L}

\item{Align}{= 'center'}

\item{DropDownRight}{= FALSE}

\item{Animate}{= TRUE}

\item{Status}{= 'custom'}

\item{H3Color}{= '#0088a7'}
}
\description{
Display a plot dropdown button with a formatted input box for YVar, XVar, GroupVars, Levels1, Levels2, Levels3, FacetVar1, FacetVar2, SizeVar
}
\seealso{
Other Output Panels: 
\code{\link{EDAPanels}()},
\code{\link{MLPanels}()},
\code{\link{PlotPanels}()}
}
\author{
Adrian Antico
}
\concept{Output Panels}
