% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{PlotPanels}
\alias{PlotPanels}
\title{PlotPanels}
\usage{
PlotPanels(
  id,
  Page,
  NumPlotsAvailable,
  AppWidth = 12L,
  DragulaChoices = NULL,
  PlotPanel = TRUE
)
}
\arguments{
\item{id}{= 'PlotPanels'}

\item{Page}{numeric}

\item{NumPlotsAvailable}{numeric}

\item{AppWidth}{= 12L}

\item{DragulaChoices}{character}
}
\description{
Panels
}
\seealso{
Other Output Panels: 
\code{\link{EDAPanels}()},
\code{\link{MLPanels}()},
\code{\link{PlotDropDownContents}()}
}
\author{
Adrian Antico
}
\concept{Output Panels}
