% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Plots.ParallelPlot.Modal}
\alias{Plots.ParallelPlot.Modal}
\title{Plots.ParallelPlot.Modal}
\usage{
Plots.ParallelPlot.Modal(id, PlotNumber, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotVariablesModals'}

\item{PlotNumber}{1}

\item{AppWidth}{= 12L}
}
\description{
Distribution modals. Histogram, Density, River, and Pie Charts, where only Y and GroupVars are needed
}
\seealso{
Other Plotting Modals: 
\code{\link{DataFilters}()},
\code{\link{Plots.3D.Modal.NoFacet}()},
\code{\link{Plots.3D.Modal}()},
\code{\link{Plots.3D.NoGroup.Modal}()},
\code{\link{Plots.3D.Target.NoGroup.Modal}()},
\code{\link{Plots.ACF.Modal}()},
\code{\link{Plots.AggRel.Modal2}()},
\code{\link{Plots.AggRel.Modal}()},
\code{\link{Plots.AggRel.NoFacet.Modal}()},
\code{\link{Plots.CorrMatrix.Modal}()},
\code{\link{Plots.Distribution.Modal}()},
\code{\link{Plots.Distribution.XY.NoFacet.Modal}()},
\code{\link{Plots.Distribution.Y.NoFacet.Modal}()},
\code{\link{Plots.Eval.Modal}()},
\code{\link{Plots.Eval.ShapImp.Modal}()},
\code{\link{Plots.Eval.VarImp.Modal}()},
\code{\link{Plots.Eval.XY.Modal}()},
\code{\link{Plots.Eval.XYZ.NoGroup.Modal}()},
\code{\link{Plots.Eval.ZY.Modal}()},
\code{\link{Plots.Eval.ZY.NoFacet.Modal}()},
\code{\link{Plots.ProbabilityPlot.Modal}()},
\code{\link{Plots.Radar.Modal}()},
\code{\link{Plots.WordCloud.Modal}()}
}
\author{
Adrian Antico
}
\concept{Plotting Modals}
