% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{PlottingReport}
\alias{PlottingReport}
\title{PlottingReport}
\usage{
PlottingReport(PlotOutputList = NULL, OutputPath = NULL)
}
\arguments{
\item{PlotOutputList}{NULL}

\item{OutputPath}{List of output objects}
}
\description{
PlottingReport is an Rmarkdown report for viewing Plotting results
}
\seealso{
Other Reports: 
\code{\link{EDAReport}()},
\code{\link{ModelInsightsReport}()},
\code{\link{TablesReport}()}
}
\author{
Adrian Antico
}
\concept{Reports}
