% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{PostGRE_Modal_Fun}
\alias{PostGRE_Modal_Fun}
\title{PostGRE_Modal_Fun}
\usage{
PostGRE_Modal_Fun(
  id,
  AppWidth = 12L,
  PostGRE_PullTable_Table_Selected = NULL,
  PostGRE_PullTable_Columns_Selected = NULL,
  PostGRE_PullTable_GroupByColumns_Selected = NULL,
  PostGRE_DropTable_Table_Selected = NULL,
  PostGRE_PullTable_AggStat_Choices = NULL,
  PostGRE_SamplePercent_Choices = NULL,
  PostGRE_PullTable_DB_Choices = NULL,
  PostGRE_DropTable_DB_Choices = NULL,
  PostGRE_DropDB_DB_Choices = NULL,
  SaveData_SelectDataPostGRE_Choices = NULL,
  SaveData_DataBaseName_Choices_Choices = NULL,
  PostGRE_PullTable_AggStat_Selected = NULL,
  PostGRE_SamplePercent_Selected = NULL,
  PostGRE_PullTable_DB_Selected = NULL,
  PostGRE_DropTable_DB_Selected = NULL,
  PostGRE_NewDB_DB_Selected = NULL,
  PostGRE_DropDB_DB_Selected = NULL,
  SaveData_SelectDataPostGRE_Selected = NULL,
  SaveData_DataBaseName_Choices_Selected = NULL,
  SaveData_TableName_Selected = NULL
)
}
\arguments{
\item{id}{= 'PostGREDMID'}

\item{AppWidth}{= 12L}
}
\description{
PostGRE Environment
}
\seealso{
Other DM: 
\code{\link{AzureBlob_Modal_Fun}()},
\code{\link{DM_AzureBlob}()},
\code{\link{DM_Local}()},
\code{\link{DM_PostGRE}()},
\code{\link{DM_Session}()},
\code{\link{Local_Modal_Fun}()},
\code{\link{Session_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DM}
