% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{PredictionIntervals}
\alias{PredictionIntervals}
\title{EffectsInterval}
\usage{
PredictionIntervals(Model)
}
\arguments{
\item{Model}{Output from AutoQuant:: supervised learning functions}
}
\description{
Random Effects Prediction Intervals: random effect, upper and lower 95th Percentile. For various Model predictions, possibly with new data
}
\seealso{
Other Mixed Effects: 
\code{\link{EffectsInterval}()},
\code{\link{FastLM}()},
\code{\link{LMER}()},
\code{\link{MEOW}()},
\code{\link{REsim}()}
}
\author{
Adrian Antico
}
\concept{Mixed Effects}
