% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDM.R
\name{QueryBuilder}
\alias{QueryBuilder}
\title{QueryBuilder}
\usage{
QueryBuilder(
  DB = NULL,
  Table = NULL,
  Columns = NULL,
  GroupByColumns = NULL,
  AggStat = NULL,
  SamplePercent = 1,
  Limit = NULL
)
}
\arguments{
\item{DB}{value}

\item{Table}{value}

\item{Columns}{value}

\item{GroupByColumns}{value}

\item{AggStat}{"AVE", "COUNT", "MAX", "MIN", "SUM"}

\item{Limit}{10}

\item{FilterVariable1}{value}

\item{FilterVariable2}{value}

\item{FilterVariable3}{value}

\item{FilterVariable4}{value}

\item{FilterLogic1}{value}

\item{FilterLogic2}{value}

\item{FilterLogic3}{value}

\item{FilterLogic4}{value}

\item{FilterValue1}{value}

\item{FilterValue2}{value}

\item{FilterValue3}{value}

\item{FilterValue4}{value}
}
\value{
Query string
}
\description{
QueryBuilder will build a query string that can be executed
}
\examples{
\dontrun{
DB = 'AutoQuant'
Table = "BenchmarkData1.csv"
Columns = c("Region", "Store", "Dept", "Weekly_Sales")
GroupByColumns = c("Region", "Store", "Dept")
AggStat = c("AVG")
Limit = 10
FilterVariable1 = 'Region'
FilterVariable2 = 'Store'
FilterVariable3 = 'Dept'
FilterVariable4 = 'Weekly_Sales'
FilterLogic1 = 'in'
FilterLogic2 = 'in'
FilterLogic3 = 'in'
FilterLogic4 = '<'
FilterValue1 = 'A'
FilterValue2 = '1'
FilterValue3 = '1'
FilterValue4 = '5000'
}

}
\seealso{
Other Database: 
\code{\link{DM.pgAppend}()},
\code{\link{DM.pgCreateDB}()},
\code{\link{DM.pgCreateTable}()},
\code{\link{DM.pgDropDB}()},
\code{\link{DM.pgListDatabases}()},
\code{\link{DM.pgListTables}()},
\code{\link{DM.pgQuery}()},
\code{\link{DM.pgRemoveCreateAppend}()},
\code{\link{DM.pgRemoveTable}()},
\code{\link{DM.pgTableColnames}()}
}
\author{
Adrian Antico
}
\concept{Database}
