% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{REsim}
\alias{REsim}
\title{REsim}
\usage{
REsim(Model, n.sims = 200, oddsRatio = FALSE, seed = NULL)
}
\arguments{
\item{Model}{Model <- lme4::lmer(gpa ~ occasion + (1 | student), data = gpa)}

\item{n.sims}{= 200}

\item{oddsRatio}{= FALSE}

\item{seed}{= NULL}
}
\description{
Mixture modeling random effects simulation data generator. Rebuild of merTools::REsim. 8x speedup.
}
\examples{
\dontrun{

# Load Data
# https://m-clark.github.io/mixed-models-with-R/random_intercepts.html#fnref10
gpa <- data.table::fread(file = "C:/Users/Bizon/Documents/GitHub/gpa.csv")
Model = lme4::lmer(gpa ~ occasion + (1 | student), data = gpa)
summary(Model)

# Code
# paste0(as.character(Model@call), collapse = " ")
head(lme4::ranef(Model),5)
mixedup::extract_random_effects(Model)

microbenchmark::microbenchmark(
  times = 10,
  merTools::REsim(Model),
  DataMuse::REsim(Model))
}

}
\seealso{
Other Mixed Effects: 
\code{\link{EffectsInterval}()},
\code{\link{FastLM}()},
\code{\link{LMER}()},
\code{\link{MEOW}()},
\code{\link{PredictionIntervals}()}
}
\author{
Adrian Antico
}
\concept{Mixed Effects}
