% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ReturnParam}
\alias{ReturnParam}
\title{Save VarName values within a project}
\usage{
ReturnParam(
  xx = NULL,
  VarName = NULL,
  Type = "numeric",
  Default = NULL,
  Switch = TRUE,
  Debug = FALSE
)
}
\arguments{
\item{xx}{This is the input value within a Shiny context; input$value or better yet, tryCatch({input$value}, error = function(x) NULL)}

\item{VarName}{The name of the argument you want to store}

\item{Type}{'character' 'numeric' 'logical' 'date'}

\item{Default}{default value}

\item{Switch}{= FALSE}

\item{Debug}{= FALSE}
}
\value{
Updates ProjectList inside function
}
\description{
Automatically save VarNames to project list
}
\examples{
\dontrun{
Aggregate <- DataMuse:::ReturnParam(input, VarName = "TS_AggregateFunction", Type = "character", Default = "mean")
}

}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
