% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{rgba2hex}
\alias{rgba2hex}
\title{rgba2hex}
\usage{
rgba2hex(Color)
}
\arguments{
\item{Color}{hex or rgba value}
}
\value{
hex value
}
\description{
rgba2hex
}
\seealso{
Other Graphics: 
\code{\link{Shiny.Display}()},
\code{\link{Shiny.Plot.Build}()},
\code{\link{Shiny.Plot.DM.data1}()},
\code{\link{Shiny.Plot.StandardPlotChecks}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
\keyword{internal}
