% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{runQuantico}
\alias{runQuantico}
\title{runQuantico}
\usage{
runQuantico(
  UserName_Password_DT = NULL,
  MaxTabs = 3L,
  AzureCredsFile = NULL,
  WorkingDirectory = getwd(),
  PostGRE_DBNames = NULL,
  PostGRE_Host = NULL,
  PostGRE_Port = NULL,
  PostGRE_User = NULL,
  PostGRE_Password = NULL,
  Debug = FALSE,
  DebugSessionLoad = FALSE,
  DebugAddPlotTab = FALSE,
  DebugPlotButtonTextUpdate = FALSE,
  DebugAddPlotButton = FALSE,
  DebugPlottingOE = FALSE,
  DebugFC = FALSE
)
}
\arguments{
\item{UserName_Password_DT}{NULL. In order to enforce authentication, supply a data.table with columns 'UserName' which contains the names of your users and 'Password' which contains the acceptable passwords. E.g. data.table::data.table(UserName = c('Adrian Antico', 'Guest'), Password = c('Password1', 'Password2')). Case sensitivity applies.}

\item{MaxTabs}{10L}

\item{AzureCredsFile}{Path to your credentials file}

\item{WorkingDirectory}{getwd() or your preferred location. Saving local data will output to this directory}

\item{PostGRE_DBNames}{Names of warehouses you want available from your postgres database}

\item{PostGRE_Host}{Host for your postgres database}

\item{PostGRE_Port}{Port for your postgres database}

\item{PostGRE_User}{Username for your postgres database}

\item{PostGRE_Password}{Password for your postgres database}

\item{Debug}{FALSE}

\item{DebugSessionLoad}{FALSE}

\item{DebugAddPlotTab}{FALSE}

\item{DebugPlotButtonTextUpdate}{FALSE}

\item{DebugAddPlotButton}{FALSE}

\item{DebugPlottingOE}{FALSE}

\item{DebugFC}{Debugging for Forecasting}
}
\description{
Shiny web app with capabilities in data managment, data wrangling, feature engineering, machine learning, statistical modeling, forecasting, visualization, and code print outs in multiple languages.
}
\examples{
\dontrun{
# Pull Data
# data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "AutoQuant"))
# data[, Date := as.Date(Date)]

# Creds file should look like this in data.table form. Two columns, Account and Values.

# Account                Values
# StorageAccount         rshinyapps
# Container              blobstorageaccess
# Key                    ...
# PostGRE_User           postgres
# PostGRE_Password       ....
# PostGRE_Host           local
# PostGRE_Port           5432

# Run App
Quantico::rapp()

}

}
\author{
Adrian Antico
}
\concept{GUI}
