% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Session_Modal_Fun}
\alias{Session_Modal_Fun}
\title{Local_Modal_Fun}
\usage{
Session_Modal_Fun(
  id,
  AppWidth = 12L,
  LoadSessionName_Choices = NULL,
  LoadSession_IncludeData_Choices = NULL,
  SaveSession_IncludeData_Choices = NULL,
  LoadSessionName_Selected = NULL,
  SaveSessionName_Selected = NULL,
  LoadSession_IncludeData_Selected = NULL,
  SaveSession_IncludeData_Selected = NULL
)
}
\arguments{
\item{id}{= 'SessionDMID'}

\item{AppWidth}{= 12L}
}
\description{
Local Environment
}
\seealso{
Other DM: 
\code{\link{AzureBlob_Modal_Fun}()},
\code{\link{DM_AzureBlob}()},
\code{\link{DM_Local}()},
\code{\link{DM_PostGRE}()},
\code{\link{DM_Session}()},
\code{\link{Local_Modal_Fun}()},
\code{\link{PostGRE_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{DM}
