% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Shapiro.Test}
\alias{Shapiro.Test}
\title{Shapiro.Test}
\usage{
Shapiro.Test(Vals, SampleSize = 5000, Samples = 30)
}
\arguments{
\item{Vals}{a vector of values to test}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Shapiro.Test
}
\examples{
\dontrun{
output <- Shapiro.Test(Vals = runif(1000), SampleSize = 100, Samples = 500)
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()}
}
\concept{Inference}
