% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUtils.R
\name{Shiny.CodePrint.Collect}
\alias{Shiny.CodePrint.Collect}
\title{Shiny.CodePrint.Collect}
\usage{
Shiny.CodePrint.Collect(y, x)
}
\arguments{
\item{y}{collection list to append to}

\item{x}{code being collected}

\item{Debug}{passthrough}
}
\description{
Collect code
}
\seealso{
Other Utils: 
\code{\link{CleanVector}()},
\code{\link{DataTable}()},
\code{\link{IntraSessionDefaults}()},
\code{\link{Shiny.CodePrint.OrganizeCode}()},
\code{\link{Shiny.Utils.CachePath}()},
\code{\link{Utils.Code.Lines}()},
\code{\link{Utils.LinuxRunSH}()},
\code{\link{Utils.Nest}()}
}
\author{
Adrian Antico
}
\concept{Utils}
