% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Shiny.Display}
\alias{Shiny.Display}
\title{Shiny.Display}
\usage{
Shiny.Display(
  input,
  output,
  session,
  PlotList,
  Debug,
  OutputId,
  Cols = 1,
  FontColor = "white",
  PM = NULL
)
}
\arguments{
\item{input}{From App}

\item{output}{From App}

\item{session}{From App}

\item{PlotList}{Named list \code{list("Plot 1" = "Plot 1", "Updated Name" = "Updated Name", ...)} containing names of plot button objects}

\item{Debug}{From App}

\item{OutputId}{From App}

\item{Cols}{Multiplot output number of columns. Defaults to 1}

\item{FontColor}{white}

\item{PM}{= NULL}
}
\description{
Output display of plots for shiny apps.
}
\details{
Output as many plots as you desire. Mix and match plotting backends. Alter sizing, spacing, and styling.
}
\seealso{
Other Graphics: 
\code{\link{Shiny.Plot.Build}()},
\code{\link{Shiny.Plot.DM.data1}()},
\code{\link{Shiny.Plot.StandardPlotChecks}()},
\code{\link{rgba2hex}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
\keyword{internal}
