% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.CARMA}
\alias{Shiny.FC.CARMA}
\title{Shiny.FC.CARMA}
\usage{
Shiny.FC.CARMA(
  input,
  output,
  DataList,
  ArgsList,
  CodeList,
  DebugFC,
  TabCount = 5L,
  Algo = "CatBoost",
  wd = NULL
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{DataList}{DataList stores data in app}

\item{ArgsList}{ArgsList}

\item{CodeList}{CodeList from app}

\item{DebugFC}{DebugFC from app}

\item{Algo}{'CatBoost', 'XGBoost', 'LightGBM', 'H2O_DRF', 'H2O_GBM'}
}
\value{
a list of columns names by data type
}
\description{
Shiny forecasting function
}
\details{
# There are 5 run modes for this function

-'Train New Model': prepare for backtesting / forecasting; inspect ML performance

-'Retrain Existing Model': Load model, new data, but don't need forecast yet

-'Backtest': evaluate model in variety of ways

-'Forecast': generate a forecast from a trained model; either loaded or created in session

-'Forecast+Retrain': generate an updated model and use that model to generate a forecast
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
