% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.Panel.Backest.FeatureEval}
\alias{Shiny.FC.Panel.Backest.FeatureEval}
\title{Shiny.FC.Panel.Backest.FeatureEval}
\usage{
Shiny.FC.Panel.Backest.FeatureEval(
  LoopMetrics,
  ArgsList,
  CodeList,
  DataList,
  ModelID,
  Algo = "catboost",
  VD,
  DebugFC = FALSE
)
}
\arguments{
\item{LoopMetrics}{From function above}

\item{ArgsList}{List}

\item{CodeList}{From app}

\item{DataList}{From app}

\item{ModelID}{From Panel Forecasting Common Args Section of Shiny.FC.Trainer}

\item{Algo}{= 'catboost', 'xgboost', 'lightgbm'}

\item{VD}{ValidationData}

\item{DebugFC}{From app}

\item{Rollup}{= FALSE. If TRUE, a rollup is made from the _Raw data}
}
\description{
Test out various combinations of structure parameters to narrow the search space
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.ReportOutput}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
