% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.Panel.Backest.RollingEval}
\alias{Shiny.FC.Panel.Backest.RollingEval}
\title{Shiny.FC.Panel.Backest.RollingEval}
\usage{
Shiny.FC.Panel.Backest.RollingEval(
  ArgsList,
  DataList,
  CodeList,
  ModelID,
  VD = NULL,
  mu = TRUE,
  vu = TRUE,
  vuf = 1,
  muf = 1,
  Algo = "catboost",
  DebugFC = FALSE
)
}
\arguments{
\item{ArgsList}{List}

\item{DataList}{From app}

\item{CodeList}{From app}

\item{ModelID}{From Panel Forecasting Common Args Section of Shiny.FC.Trainer}

\item{VD}{ValidationData}

\item{mu}{Model updating}

\item{vu}{Values updating}

\item{vuf}{Value update frequency = every nth FC_Period. E.g., update data daily vs weekly}

\item{muf}{Model update frequency = every nth FC_Period}

\item{Algo}{= 'catboost', 'xgboost', 'lightgbm'}

\item{DebugFC}{From app}
}
\description{
Goal of this function is to get a better understanding of the cost tradeoffs for different model configurations
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
