% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.Panel.Metrics.Agg}
\alias{Shiny.FC.Panel.Metrics.Agg}
\title{Shiny.FC.Panel.Metrics.Agg}
\usage{
Shiny.FC.Panel.Metrics.Agg(
  DataList,
  ArgsList,
  ModelID,
  GroupVariables = NULL,
  CrossEval = FALSE,
  Metrics = c("AvgError", "AccumError", "RMSE", "MAE", "MAPE", "SMAPE"),
  DebugFC = FALSE
)
}
\arguments{
\item{DataList}{From App}

\item{ArgsList}{From App}

\item{ModelID}{From App}

\item{GroupVariables}{From App}

\item{CrossEval}{= FALSE. Set to TRUE for CrossEval agg}

\item{Metrics}{From App}

\item{DebugFC}{= FALSE}

\item{Rollup}{= FALSE. Set to TRUE to get an agg set of eval metrics}
}
\description{
Aggregate Raw Metrics, update DataList
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS.ReportOutput}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
