% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.Panel.Metrics.Raw}
\alias{Shiny.FC.Panel.Metrics.Raw}
\title{Shiny.FC.Panel.Metrics.Raw}
\usage{
Shiny.FC.Panel.Metrics.Raw(
  DataList,
  ArgsList,
  ModelID,
  GroupVariables = NULL,
  Metrics = c("AvgError", "AccumError", "RMSE", "MAE", "MAPE", "SMAPE"),
  CrossEval = FALSE,
  DebugFC = FALSE
)
}
\arguments{
\item{DataList}{From App}

\item{ArgsList}{From App}

\item{ModelID}{From App}

\item{GroupVariables}{From App}

\item{Metrics}{'AvgError', 'AccumError', 'RMSE', 'MAE', 'MAPE', 'SMAPE'}

\item{DebugFC}{= FALSE}
}
\description{
Forecasting create error metrics columns
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
