% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.FC.Panel.Train}
\alias{Shiny.FC.Panel.Train}
\title{Shiny.FC.Panel.Train}
\usage{
Shiny.FC.Panel.Train(
  ArgsList,
  CodeList,
  DataList,
  ModelID,
  Algo = "CatBoost",
  VD = NULL,
  DebugFC = FALSE
)
}
\arguments{
\item{ArgsList}{List}

\item{CodeList}{From app}

\item{DataList}{From app}

\item{ModelID}{From Panel Forecasting Common Args Section of Shiny.FC.Trainer}

\item{Algo}{'CatBoost', 'XGBoost', 'LightGBM'}

\item{VD}{ValidationData}

\item{DebugFC}{From app}
}
\description{
Step 1 in the forecasting process, train the model
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
