% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFC.R
\name{Shiny.FC.SS.ReportOutput}
\alias{Shiny.FC.SS.ReportOutput}
\title{Shiny.FC.SS.ReportOutput}
\usage{
Shiny.FC.SS.ReportOutput(
  input,
  output,
  DataList,
  CodeList,
  Page,
  Debug = FALSE,
  MOL = NULL,
  Algo = NULL,
  ModelID = NULL,
  RunMode = NULL,
  Theme = "dark",
  FontColor = NULL,
  PlotWidth = "1450px",
  PlotHeight = "850px"
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{DataList}{DataList stores data in app}

\item{CodeList}{CodeList from app}

\item{Debug}{DebugFC from app}

\item{ArgsList}{ArgsList}
}
\value{
a list of columns names by data type
}
\description{
Shiny FC SS Report
}
\seealso{
Other FC: 
\code{\link{BasicLoop.SingleTest}()},
\code{\link{FC.Backtest.CalendarGroups}()},
\code{\link{FC.Backtest.DateGroups}()},
\code{\link{FC.DateCast}()},
\code{\link{FC.FCPeriods}()},
\code{\link{FC.TimeUnit}()},
\code{\link{FC.XREGS}()},
\code{\link{Shiny.EDA.ReportOutput}()},
\code{\link{Shiny.FC.CARMA}()},
\code{\link{Shiny.FC.Panel.Backest.FeatureEval}()},
\code{\link{Shiny.FC.Panel.Backest.RollingEval}()},
\code{\link{Shiny.FC.Panel.Backtest}()},
\code{\link{Shiny.FC.Panel.Forecast}()},
\code{\link{Shiny.FC.Panel.Metric.Rollup}()},
\code{\link{Shiny.FC.Panel.Metrics.Agg}()},
\code{\link{Shiny.FC.Panel.Metrics.Raw}()},
\code{\link{Shiny.FC.Panel.ReportOutput}()},
\code{\link{Shiny.FC.Panel.Retrain}()},
\code{\link{Shiny.FC.Panel.Train}()},
\code{\link{Shiny.FC.SS}()},
\code{\link{Shiny.ML.ReportOutput}()}
}
\author{
Adrian Antico
}
\concept{FC}
