% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Shiny.Plot.Build}
\alias{Shiny.Plot.Build}
\title{Shiny.Plot.Build}
\usage{
Shiny.Plot.Build(
  input,
  output,
  session,
  PlotNumbers,
  DataList,
  Debug,
  PlotMap,
  CodeList,
  EchartsTheme,
  OutputPage = 1,
  PlotEngine = "Plotly",
  Timeline = FALSE,
  MouseScroll = FALSE,
  ColorFont = "#e2e2e2",
  NumLevelsDisplay = 50L,
  NumberOfBins = 30L,
  GlobalChange = FALSE,
  PlotHeight = "860px",
  PlotWidth = "1450px",
  Args = NULL,
  SubsetList = NULL
)
}
\arguments{
\item{input}{From App}

\item{output}{From App}

\item{session}{From App}

\item{DataList}{From App}

\item{Debug}{From App}

\item{CodeList}{From App}

\item{EchartsTheme}{= EchartsTheme From App}

\item{OutputPage}{From App}

\item{PlotEngine}{= PE from App}

\item{Timeline}{= TL from App}

\item{ColorFont}{= "#e2e2e2"}

\item{NumLevelsDisplay}{= numeric}

\item{NumberOfBins}{histograms}

\item{GlobalChange}{TRUE}

\item{Args}{NULL}

\item{SubsetList}{NULL}

\item{PlotNumberss}{From App}
}
\description{
Step 1 in building plots
}
\seealso{
Other Graphics: 
\code{\link{Shiny.Display}()},
\code{\link{Shiny.Plot.DM.data1}()},
\code{\link{Shiny.Plot.StandardPlotChecks}()},
\code{\link{rgba2hex}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
\keyword{internal}
