% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Shiny.Plot.StandardPlotChecks}
\alias{Shiny.Plot.StandardPlotChecks}
\title{Shiny.Plot.StandardPlotChecks}
\usage{
Shiny.Plot.StandardPlotChecks(
  d = NULL,
  y = NULL,
  x = NULL,
  z = NULL,
  Type = NULL
)
}
\arguments{
\item{d}{From App}

\item{y}{From App}

\item{x}{From App}

\item{z}{From App}

\item{Type}{From App}
}
\description{
Ensure variables are defined correctly for standard plots, if applicable
}
\seealso{
Other Graphics: 
\code{\link{Shiny.Display}()},
\code{\link{Shiny.Plot.Build}()},
\code{\link{Shiny.Plot.DM.data1}()},
\code{\link{rgba2hex}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
\keyword{internal}
