% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{Shrink_Modal_Fun}
\alias{Shrink_Modal_Fun}
\title{Shrink_Modal_Fun}
\usage{
Shrink_Modal_Fun(
  id,
  AppWidth = 12L,
  Aggregate_Columns_Selected = NULL,
  Aggregate_ByVariables_Selected = NULL,
  Aggregate_DateVariable_Selected = NULL,
  SubsetData_FilterVariable1_Selected = NULL,
  SubsetData_FilterVariable2_Selected = NULL,
  SubsetData_FilterVariable3_Selected = NULL,
  SubsetData_FilterVariable4_Selected = NULL,
  SubsetData_FilterValue_1_1_Selected = NULL,
  SubsetData_FilterValue_1_2_Selected = NULL,
  SubsetData_FilterValue_2_1_Selected = NULL,
  SubsetData_FilterValue_2_2_Selected = NULL,
  SubsetData_FilterValue_3_1_Selected = NULL,
  SubsetData_FilterValue_3_2_Selected = NULL,
  SubsetData_FilterValue_4_1_Selected = NULL,
  SubsetData_FilterValue_4_2_Selected = NULL,
  DeleteVariables_Selected = NULL,
  SampleData_StratifyColumns_Selected = NULL,
  AggregateData_SelectData_Choices = NULL,
  Aggregate_Stat_Choices = NULL,
  Aggregate_TimeAgg_Choices = NULL,
  SubsetData_FilterLogic1_Choices = NULL,
  SubsetData_FilterLogic2_Choices = NULL,
  SubsetData_FilterLogic3_Choices = NULL,
  SubsetData_FilterLogic4_Choices = NULL,
  SubsetData_SelectData_Choices = NULL,
  DeleteVariables_SelectData_Choices = NULL,
  SampleData_SelectData_Choices = NULL,
  SampleData_Rate_Choices = NULL,
  AggregateData_SelectData_Selected = NULL,
  AggregateData_NewName_Selected = NULL,
  Aggregate_Stat_Selected = NULL,
  Aggregate_TimeAgg_Selected = NULL,
  SubsetData_FilterLogic1_Selected = NULL,
  SubsetData_FilterLogic2_Selected = NULL,
  SubsetData_FilterLogic3_Selected = NULL,
  SubsetData_FilterLogic4_Selected = NULL,
  SubsetData_SelectData_Selected = NULL,
  SubsetData_NewName_Selected = NULL,
  DeleteVariables_SelectData_Selected = NULL,
  DeleteVariables_NewName_Selected = NULL,
  SampleData_SelectData_Selected = NULL,
  SampleData_NewName_Selected = NULL,
  SampleData_Rate_Selected = NULL
)
}
\arguments{
\item{id}{= 'Shrink_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Subset, Sample, Aggregate
}
\seealso{
Other DW: 
\code{\link{Columns_Modal_Fun}()},
\code{\link{DW_Columns}()},
\code{\link{DW_GrowData}()},
\code{\link{DW_Misc}()},
\code{\link{DW_Pivot}()},
\code{\link{DW_ShrinkData}()},
\code{\link{DataSets_Modal_Fun}()},
\code{\link{Grow_Modal_Fun}()},
\code{\link{Misc_Modal_Fun}()},
\code{\link{Pivot_Modal_Fun}()},
\code{\link{Shiny.DW.AggregateData}()},
\code{\link{Shiny.DW.CastData}()},
\code{\link{Shiny.DW.ConcatenateColumns}()},
\code{\link{Shiny.DW.DeleteColumns}()},
\code{\link{Shiny.DW.JoinData}()},
\code{\link{Shiny.DW.MeltData}()},
\code{\link{Shiny.DW.MetaProgramming}()},
\code{\link{Shiny.DW.RemoveData}()},
\code{\link{Shiny.DW.RenameColumns}()},
\code{\link{Shiny.DW.SampleData}()},
\code{\link{Shiny.DW.SortData}()},
\code{\link{Shiny.DW.SubsetData}()},
\code{\link{Shiny.DW.TimeSeriesFill}()},
\code{\link{Shiny.DW.TimeSeriesRoll}()},
\code{\link{Shiny.DW.TimeTrendColumn}()},
\code{\link{Shiny.DW.TypeCast}()},
\code{\link{Shiny.DW.UnionData}()}
}
\author{
Adrian Antico
}
\concept{DW}
