% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomePage.R
\name{SideBarUI}
\alias{SideBarUI}
\title{SideBarUI}
\usage{
SideBarUI(id, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotPanels'}

\item{AppWidth}{= 12L}

\item{Page}{numeric}

\item{NumPlotsAvailable}{numeric}

\item{DragulaChoices}{character}
}
\description{
Panels
}
\seealso{
Other Plotting: 
\code{\link{CodePrintPage}()},
\code{\link{HomePage}()},
\code{\link{InferencePage}()}
}
\author{
Adrian Antico
}
\concept{Plotting}
