% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyInputs.R
\name{TextInput}
\alias{TextInput}
\title{TextInput}
\usage{
TextInput(
  InputID = NULL,
  input = NULL,
  Label = NULL,
  Value = NULL,
  PlotDropDown = NULL,
  Key = NULL,
  Width = "100\%",
  Placeholder = "",
  session = NULL,
  Update = FALSE
)
}
\arguments{
\item{InputID}{ID of the input}

\item{input}{For update usage. Provide input if you want your inputs to update with previous values when screens reload}

\item{Label}{Feeds label}

\item{Value}{Default}

\item{PlotDropDown}{From App}

\item{Key}{From App}

\item{Width}{width arg}

\item{session}{For update usgae}

\item{Update}{FALSE}

\item{server}{For update usage}

\item{Format}{Date format}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
TextInput automatically builds a text input with ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_Date <- renderUI({
  Quantico:::TextInput(InputID = "TS_CARMA_HolidayMovingAverages",
                         Label = "Import Data Creation Date",
                         Value = Sys.Date(),
                         Placeholder = "yyyy-mm-dd")})
}
}
\seealso{
Other Shiny: 
\code{\link{BlankLine}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput1}()},
\code{\link{SelectizeInput}()},
\code{\link{SliderInput}()},
\code{\link{TextAreaInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
