% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{Two.Sample.TTest}
\alias{Two.Sample.TTest}
\title{Two.Sample.TTest}
\usage{
Two.Sample.TTest(
  dt = NULL,
  Variable1 = NULL,
  Variable2 = NULL,
  Paired = FALSE,
  EqualVariance = FALSE,
  ConfidenceLevel = 0.95,
  MeanDifference = 0,
  Alternative = "two.sided",
  SampleSize = NULL,
  Samples = 1,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "850px",
  PlotWidth = "1450px"
)
}
\arguments{
\item{dt}{data.table}

\item{Paired}{logical}

\item{EqualVariance}{logical}

\item{ConfidenceLevel}{numeric}

\item{MeanDifference}{H0 value}

\item{Alternative}{"two.sided", "less", or "greater"}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}

\item{Variables1}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}

\item{Variables2}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}
}
\description{
Two.Sample.TTest
}
\examples{
\dontrun{
Variable1 = "Daily Margin"
dt = data.table::data.table(qnorm(p = runif(100000)))
data.table::setnames(dt, "V1", Variable1)
Variable2 = "Daily Liters"
dt[, eval(Variable2) := qnorm(p = runif(100000))]
Paired = FALSE
MeanDifference = 0
Alternative = "two.sided"
ConfidenceLevel = 0.95
SampleSize = 500
Samples = 50
EchartsTheme = "macarons"
TextColor = "black"
PlotHeight = "850px"
PlotWidth = "1450px"
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\concept{Inference}
