% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{TwoSampleTTest_Modal_Fun}
\alias{TwoSampleTTest_Modal_Fun}
\title{TwoSampleTTest_Modal_Fun}
\usage{
TwoSampleTTest_Modal_Fun(
  id,
  AppWidth = 12L,
  TwoSampleTTest_Variable1_Selected = NULL,
  TwoSampleTTest_Variable2_Selected = NULL,
  TwoSampleTTest_SelectData_Choices = NULL,
  TwoSampleTTest_Alternative_Choices = NULL,
  TwoSampleTTest_Paired_Choices = NULL,
  TwoSampleTTest_EqualVariance_Choices = NULL,
  TwoSampleTTest_SelectData_Selected = NULL,
  TwoSampleTTest_InferenceID_Selected = NULL,
  TwoSampleTTest_Paired_Selected = NULL,
  TwoSampleTTest_EqualVariance_Selected = NULL,
  TwoSampleTTest_NullValue_Selected = NULL,
  TwoSampleTTest_Alternative_Selected = NULL,
  TwoSampleTTest_ConfidenceLevel_Selected = NULL,
  TwoSampleTTest_Samples_Selected = NULL,
  TwoSampleTTest_SampleSize_Selected = NULL
)
}
\arguments{
\item{id}{= 'TwoSampleTTest_Modal_Fun'}

\item{AppWidth}{= 12L}
}
\description{
Inference operations
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()}
}
\author{
Adrian Antico
}
\concept{Inference}
