% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{XGBoostCARMA_Modal_Fun}
\alias{XGBoostCARMA_Modal_Fun}
\title{XGBoostCARMA_Modal_Fun}
\usage{
XGBoostCARMA_Modal_Fun(
  id,
  AppWidth = 12L,
  XGBoostCARMA_ModelID_Selected = NULL,
  XGBoostCARMA_TargetColumnName_Selected = NULL,
  XGBoostCARMA_DateColumnName_Selected = NULL,
  XGBoostCARMA_GroupVariables_Selected = NULL,
  XGBoostCARMA_RunMode_Choices = NULL,
  XGBoostCARMA_TimeUnit_Choices = NULL,
  XGBoostCARMA_FCPeriods_Choices = NULL,
  XGBoostCARMA_TaskType_Choices = NULL,
  XGBoostCARMA_NumGPU_Choices = NULL,
  XGBoostCARMA_SaveModel_Choices = NULL,
  XGBoostCARMA_ArgsList_Choices = NULL,
  XGBoostCARMA_TimeWeights_Choices = NULL,
  XGBoostCARMA_Methods_Choices = NULL,
  XGBoostCARMA_Difference_Choices = NULL,
  XGBoostCARMA_NonNegativePred_Choices = NULL,
  XGBoostCARMA_RoundPreds_Choices = NULL,
  XGBoostCARMA_CalendarVariables_Choices = NULL,
  XGBoostCARMA_HolidayVariables_Choices = NULL,
  XGBoostCARMA_Lookback_Choices = NULL,
  XGBoostCARMA_Lags_Choices = NULL,
  XGBoostCARMA_MovingAverages_Choices = NULL,
  XGBoostCARMA_MovingSD_Choices = NULL,
  XGBoostCARMA_MovingSkew_Choices = NULL,
  XGBoostCARMA_MovingKurt_Choices = NULL,
  XGBoostCARMA_MovingQuantiles_Choices = NULL,
  XGBoostCARMA_Quantiles_Selected_Choices = NULL,
  XGBoostCARMA_AnomalyDetection_HighThreshold_Choices = NULL,
  XGBoostCARMA_AnomalyDetection_LowThreshold_Choices = NULL,
  XGBoostCARMA_TimeTrend_Choices = NULL,
  XGBoostCARMA_DataTruncate_Choices = NULL,
  XGBoostCARMA_EncodingMethod_Choices = NULL,
  XGBoostCARMA_ZeroPadSeries_Choices = NULL,
  XGBoostCARMA_CrossEvalModelUpdate_Choices = NULL,
  XGBoostCARMA_CrossEvalValuesUpdate_Choices = NULL,
  XGBoostCARMA_CrossEvalModelUpdateFreq_Choices = NULL,
  XGBoostCARMA_CrossEvalValuesUpdateFreq_Choices = NULL,
  XGBoostCARMA_EvalMetric_Choices = NULL,
  XGBoostCARMA_LossFunction_Choices = NULL,
  XGBoostCARMA_TrainData_Choices = NULL,
  XGBoostCARMA_ValidationData_Choices = NULL,
  XGBoostCARMA_XREGS_Choices = NULL,
  XGBoostCARMA_NTrees_Choices = NULL,
  XGBoostCARMA_MaxDepth_Choices = NULL,
  XGBoostCARMA_LearningRate_Choices = NULL,
  XGBoostCARMA_MinChildWeight_Choices = NULL,
  XGBoostCARMA_SubSample_Choices = NULL,
  XGBoostCARMA_ColSampleByTree_Choices = NULL,
  XGBoostCARMA_lambda_Choices = NULL,
  XGBoostCARMA_alpha_Choices = NULL,
  XGBoostCARMA_RunMode_Selected = NULL,
  XGBoostCARMA_TimeUnit_Selected = NULL,
  XGBoostCARMA_FCPeriods_Selected = NULL,
  XGBoostCARMA_TaskType_Selected = NULL,
  XGBoostCARMA_NumGPU_Selected = NULL,
  XGBoostCARMA_SaveModel_Selected = NULL,
  XGBoostCARMA_ArgsList_Selected = NULL,
  XGBoostCARMA_TimeWeights_Selected = NULL,
  XGBoostCARMA_Methods_Selected = NULL,
  XGBoostCARMA_Difference_Selected = NULL,
  XGBoostCARMA_NonNegativePred_Selected = NULL,
  XGBoostCARMA_RoundPreds_Selected = NULL,
  XGBoostCARMA_CalendarVariables_Selected = NULL,
  XGBoostCARMA_HolidayVariables_Selected = NULL,
  XGBoostCARMA_Lookback_Selected = NULL,
  XGBoostCARMA_Lags_Selected = NULL,
  XGBoostCARMA_MovingAverages_Selected = NULL,
  XGBoostCARMA_MovingSD_Selected = NULL,
  XGBoostCARMA_MovingSkew_Selected = NULL,
  XGBoostCARMA_MovingKurt_Selected = NULL,
  XGBoostCARMA_MovingQuantiles_Selected = NULL,
  XGBoostCARMA_Quantiles_Selected_Selected = NULL,
  XGBoostCARMA_AnomalyDetection_HighThreshold_Selected = NULL,
  XGBoostCARMA_AnomalyDetection_LowThreshold_Selected = NULL,
  XGBoostCARMA_TimeTrend_Selected = NULL,
  XGBoostCARMA_DataTruncate_Selected = NULL,
  XGBoostCARMA_EncodingMethod_Selected = NULL,
  XGBoostCARMA_ZeroPadSeries_Selected = NULL,
  XGBoostCARMA_CrossEvalModelUpdate_Selected = NULL,
  XGBoostCARMA_CrossEvalValuesUpdate_Selected = NULL,
  XGBoostCARMA_CrossEvalModelUpdateFreq_Selected = NULL,
  XGBoostCARMA_CrossEvalValuesUpdateFreq_Selected = NULL,
  XGBoostCARMA_EvalMetric_Selected = NULL,
  XGBoostCARMA_LossFunction_Selected = NULL,
  XGBoostCARMA_TrainData_Selected = NULL,
  XGBoostCARMA_ValidationData_Selected = NULL,
  XGBoostCARMA_XREGS_Selected = NULL,
  XGBoostCARMA_NTrees_Selected = NULL,
  XGBoostCARMA_MaxDepth_Selected = NULL,
  XGBoostCARMA_LearningRate_Selected = NULL,
  XGBoostCARMA_MinChildWeight_Selected = NULL,
  XGBoostCARMA_SubSample_Selected = NULL,
  XGBoostCARMA_ColSampleByTree_Selected = NULL,
  XGBoostCARMA_lambda_Selected = NULL,
  XGBoostCARMA_alpha_Selected = NULL
)
}
\arguments{
\item{id}{= 'XGBoostCARMA_ID'}

\item{AppWidth}{= 12L}
}
\description{
XGBoost CARMA Parameters Modal with OK button to close
}
\seealso{
Other ML: 
\code{\link{ARFIMA_Modal_Fun}()},
\code{\link{CatBoostCARMA_Modal_Fun}()},
\code{\link{CatBoost_Modal_Fun}()},
\code{\link{CausalMediation_Modal_Fun}()},
\code{\link{ETS_Modal_Fun}()},
\code{\link{H2O_DRF_Modal_Fun}()},
\code{\link{H2O_GBM_Modal_Fun}()},
\code{\link{H2O_GLM_Modal_Fun}()},
\code{\link{H2O_HGLM_Modal_Fun}()},
\code{\link{LightGBMCARMA_Modal_Fun}()},
\code{\link{LightGBM_Modal_Fun}()},
\code{\link{NNET_Modal_Fun}()},
\code{\link{Sarima_Modal_Fun}()},
\code{\link{Shiny.ML.CatBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.CatBoost.GridEvalMetricsOptions}()},
\code{\link{Shiny.ML.CatBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.CatBoost_Params}()},
\code{\link{Shiny.ML.Evaluation.Table}()},
\code{\link{Shiny.ML.H2O.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.GLM.DistributionOptions}()},
\code{\link{Shiny.ML.H2O.LinkOptions}()},
\code{\link{Shiny.ML.H2O_DRF.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_DRF_Params}()},
\code{\link{Shiny.ML.H2O_GBM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GBM_Params}()},
\code{\link{Shiny.ML.H2O_GLM.EvalMetricOptions}()},
\code{\link{Shiny.ML.H2O_GLM_Params}()},
\code{\link{Shiny.ML.H2O_HGLM_Params}()},
\code{\link{Shiny.ML.LightGBM.LossFunctionOptions}()},
\code{\link{Shiny.ML.LightGBM_Params}()},
\code{\link{Shiny.ML.ModelDataObjects}()},
\code{\link{Shiny.ML.Scoring}()},
\code{\link{Shiny.ML.Trainer}()},
\code{\link{Shiny.ML.XGBoost.EvalMetricOptions}()},
\code{\link{Shiny.ML.XGBoost.LossFunctionOptions}()},
\code{\link{Shiny.ML.XGBoost_Params}()},
\code{\link{TBATS_Modal_Fun}()},
\code{\link{XGBoost_Modal_Fun}()}
}
\author{
Adrian Antico
}
\concept{ML}
