% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{Agostino.Test}
\alias{Agostino.Test}
\title{Agostino.Test}
\usage{
Agostino.Test(Vals, SampleSize = NULL, Samples = 1)
}
\arguments{
\item{Vals}{a vector of values to test}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Agostino.Test
}
\examples{
\dontrun{
output <- Agostino.Test(Vals = runif(100000), SampleSize = 1000, Samples = 500)
}

}
\seealso{
Other Inference: 
\code{\link{Anderson.Darling.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
