% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTests.R
\name{Correlation.Test}
\alias{Correlation.Test}
\title{Correlation.Test}
\usage{
Correlation.Test(
  dt = NULL,
  CorrVars = NULL,
  DateVar = NULL,
  SampleSize = 10000,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "600px",
  PlotWidth = "300px",
  P_Adjust = "holm",
  Bayesian = FALSE,
  Bayesian_Prior = "medium",
  MultiLevel = FALSE,
  Include_Factors = FALSE,
  Partial = FALSE,
  Partial_Bayesian = FALSE
)
}
\arguments{
\item{dt}{Numeric vector of values to test. Must have positive standard deviation and must be of length greater than or equal to 8}

\item{CorrVars}{character}

\item{SampleSize}{sub sampling of data}

\item{EchartsTheme}{= "macarons"}

\item{TextColor}{= "black"}

\item{PlotHeight}{= "600px"}

\item{PlotWidth}{= "300px"}

\item{P_Adjust}{Frequentist corrections, which include: default "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "somers" or "none"}

\item{Bayesian}{logical}

\item{Bayesian_Prior}{Choose from default "medium", "medium.narrow", "medium", "wide", and "ultrawide"}

\item{MultiLevel}{logical. TRUE using random effects, FALSE uses fixed effects}

\item{Include_Factors}{logical}

\item{Partial}{logical}

\item{Partial_Bayesian}{logical}
}
\description{
Correlation.Test
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Shapiro.Test}()}
}
\concept{Inference}
