% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{Kolmogorov.Smirnov.Test}
\alias{Kolmogorov.Smirnov.Test}
\title{Kolmogorov.Smirnov.Test}
\usage{
Kolmogorov.Smirnov.Test(Vals, SampleSize = NULL, Samples = 1)
}
\arguments{
\item{Vals}{a vector of values to test}

\item{SampleSize}{sub sampling of data}

\item{Samples}{number of iterations to run}
}
\description{
Kolmogorov.Smirnov.Test
}
\examples{
\dontrun{
output <- Kolmogorov.Smirnov.Test(runif(10000))
}

}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Normality.Analysis}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()}
}
\concept{Inference}
