% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceStats.R
\name{Normality.Analysis}
\alias{Normality.Analysis}
\title{All.Normality.Tests}
\usage{
Normality.Analysis(
  dt = NULL,
  YVars = NULL,
  EchartsTheme = "macarons",
  TextColor = "black",
  PlotHeight = "300px",
  PlotWidth = "600px",
  SampleSize.ADT = NULL,
  Samples.ADT = 1,
  SampleSize.CVMT = NULL,
  Samples.CVMT = 1,
  SampleSize.KST = NULL,
  Samples.KST = 1,
  SampleSize.ST = NULL,
  Samples.ST = 1,
  SampleSize.JBT = NULL,
  Samples.JBT = 1,
  SampleSize.AT = NULL,
  Samples.AT = 1,
  Debug = FALSE
)
}
\description{
All.Normality.Tests
}
\seealso{
Other Inference: 
\code{\link{Agostino.Test}()},
\code{\link{Anderson.Darling.Test}()},
\code{\link{ChiSq.Test}()},
\code{\link{ChiSquareTest_Modal_Fun}()},
\code{\link{Correlation.Analysis}()},
\code{\link{Correlation_Modal_Fun}()},
\code{\link{Cramer.Von.Mises.Test}()},
\code{\link{F.Test}()},
\code{\link{FTest_Modal_Fun}()},
\code{\link{Inference_1STTest}()},
\code{\link{Inference_2STTest}()},
\code{\link{Inference_ChiSq}()},
\code{\link{Inference_Correlation}()},
\code{\link{Inference_FTest}()},
\code{\link{Inference_Normality}()},
\code{\link{Jarque.Bera.Test}()},
\code{\link{Kolmogorov.Smirnov.Test}()},
\code{\link{Normality_Modal_Fun}()},
\code{\link{One.Sample.TTest}()},
\code{\link{OneSampleTTest_Modal_Fun}()},
\code{\link{Shapiro.Test}()},
\code{\link{Two.Sample.TTest}()},
\code{\link{TwoSampleTTest_Modal_Fun}()}
}
\concept{Inference}
