% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{TablesReport}
\alias{TablesReport}
\title{TablesReport}
\usage{
TablesReport(DataOutputList = NULL, OutputPath = NULL)
}
\arguments{
\item{DataOutputList}{NULL}

\item{OutputPath}{List of output objects}
}
\description{
TablesReport is an Rmarkdown report for viewing Tables results
}
\seealso{
Other Reports: 
\code{\link{EDAReport}()},
\code{\link{ModelInsightsReport}()},
\code{\link{PlottingReport}()}
}
\author{
Adrian Antico
}
\concept{Reports}
